% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-data.R
\docType{class}
\name{array-data}
\alias{array-data}
\alias{ArrayData}
\title{class ArrayData}
\description{
class ArrayData
}
\section{Usage}{
\preformatted{data <- Array$create(x)$data()

data$type()
data$length()
data$null_count()
data$offset()
data$buffers()
}
}

\section{Methods}{


...
}

\keyword{datasets}
