% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-data.R
\docType{class}
\name{ArrayData}
\alias{ArrayData}
\title{ArrayData class}
\description{
The \code{ArrayData} class allows you to get and inspect the data
inside an \code{arrow::Array}.
}
\section{Usage}{
\preformatted{data <- Array$create(x)$data()

data$type()
data$length()
data$null_count()
data$offset()
data$buffers()
}
}

\section{Methods}{


...
}

\keyword{datasets}
