% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-package.R
\docType{package}
\name{arrow-package}
\alias{arrow}
\alias{arrow-package}
\title{arrow: Integration to 'Apache' 'Arrow'}
\description{
'Apache' 'Arrow' \url{https://arrow.apache.org/} is a cross-language development platform for in-memory data. It specifies a standardized language-independent columnar memory format for flat and hierarchical data, organized for efficient analytic operations on modern hardware. This package provides an interface to the 'Arrow C++' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apache/arrow/}
  \item \url{https://arrow.apache.org/docs/r/}
  \item Report bugs at \url{https://github.com/apache/arrow/issues}
}

}
\author{
\strong{Maintainer}: Nic Crane \email{thisisnic@gmail.com}

Authors:
\itemize{
  \item Neal Richardson \email{neal.p.richardson@gmail.com}
  \item Ian Cook \email{ianmcook@gmail.com}
  \item Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})
  \item Romain François \email{romain@rstudio.com} (\href{https://orcid.org/0000-0002-2444-4226}{ORCID})
  \item Jonathan Keane \email{jkeane@gmail.com}
  \item Dragoș Moldovan-Grünfeld \email{dragos.mold@gmail.com}
  \item Jeroen Ooms \email{jeroen@berkeley.edu}
  \item Apache Arrow \email{dev@arrow.apache.org} [copyright holder]
}

Other contributors:
\itemize{
  \item Javier Luraschi \email{javier@rstudio.com} [contributor]
  \item Karl Dunkle Werner \email{karldw@users.noreply.github.com} (\href{https://orcid.org/0000-0003-0523-7309}{ORCID}) [contributor]
  \item Jeffrey Wong \email{jeffreyw@netflix.com} [contributor]
}

}
\keyword{internal}
