% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayData.R
\docType{class}
\name{arrow__ArrayData}
\alias{arrow__ArrayData}
\alias{arrow::ArrayData}
\title{class arrow::ArrayData}
\description{
class arrow::ArrayData
}
\section{Usage}{
\preformatted{data <- array(x)$data()

data$type()
data$length()
data$null_count()
data$offset()
data$buffers()
}
}

\section{Methods}{


...
}

\keyword{datasets}
