% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\docType{class}
\name{arrow__Buffer}
\alias{arrow__Buffer}
\alias{arrow::Buffer}
\title{class arrow::Buffer}
\description{
class arrow::Buffer
}
\section{Methods}{

\itemize{
\item \code{$is_mutable()} :
\item \code{$ZeroPadding()} :
\item \code{$size()} :
\item \code{$capacity()}:
}
}

\keyword{datasets}
