% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{class}
\name{arrow__DictionaryType}
\alias{arrow__DictionaryType}
\alias{arrow::DictionaryType}
\title{class arrow::DictionaryType}
\description{
class arrow::DictionaryType
}
\section{Methods}{


TODO
}

\keyword{datasets}
