% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{class}
\name{arrow__FixedWidthType}
\alias{arrow__FixedWidthType}
\alias{arrow::FixedWidthType}
\title{class arrow::FixedWidthType}
\description{
class arrow::FixedWidthType
}
\section{Methods}{


TODO
}

\keyword{datasets}
