% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{arrow__io__OutputStream}
\alias{arrow__io__OutputStream}
\alias{arrow::io::OutputStream}
\title{OutputStream}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
arrow::io::OutputStream
}
\description{
OutputStream
}
\section{Methods}{

\itemize{
\item \code{arrow::Buffer} \code{Read}(\code{int} nbytes):  Read \code{nbytes} bytes
\item \code{void} \code{close}(): close the stream
}
}

\keyword{datasets}
