% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\docType{class}
\name{arrow__ipc__Message}
\alias{arrow__ipc__Message}
\alias{arrow::ipc::Message}
\title{class arrow::ipc::Message}
\description{
class arrow::ipc::Message
}
\section{Methods}{


TODO
}

\keyword{datasets}
