% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\docType{class}
\name{arrow__ipc__MessageReader}
\alias{arrow__ipc__MessageReader}
\alias{arrow::MessageReader}
\title{class arrow::MessageReader}
\description{
class arrow::MessageReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
