% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\docType{class}
\name{arrow__ipc__RecordBatchFileReader}
\alias{arrow__ipc__RecordBatchFileReader}
\alias{arrow::RecordBatchFileReader}
\title{class arrow::RecordBatchFileReader}
\description{
class arrow::RecordBatchFileReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
