% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\docType{class}
\name{arrow__ipc__RecordBatchFileWriter}
\alias{arrow__ipc__RecordBatchFileWriter}
\alias{arrow::RecordBatchFileWriter}
\title{class arrow::RecordBatchFileWriter

Writer for the Arrow binary file format}
\description{
class arrow::RecordBatchFileWriter

Writer for the Arrow binary file format
}
\section{usage}{
\preformatted{writer <- RecordBatchFileWriter(sink, schema)

writer$write_batch(batch)
writer$write_table(table)
writer$close()
}
}

\section{Factory}{


The \code{\link[=RecordBatchFileWriter]{RecordBatchFileWriter()}} function creates a record batch stream writer.
}

\section{Methods}{

inherited from \link[=arrow__ipc__RecordBatchWriter]{arrow::RecordBatchWriter}
\itemize{
\item \code{$write_batch(batch)}: Write record batch to stream
\item \code{$write_table(table)}: write Table to stream
\item \code{$close()}: close stream
}
}

\keyword{datasets}
