% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\docType{class}
\name{arrow__ipc__RecordBatchStreamReader}
\alias{arrow__ipc__RecordBatchStreamReader}
\alias{arrow::RecordBatchStreamReader}
\title{class arrow::RecordBatchStreamReader}
\description{
class arrow::RecordBatchStreamReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
