% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\docType{class}
\name{arrow__ipc__RecordBatchStreamWriter}
\alias{arrow__ipc__RecordBatchStreamWriter}
\alias{arrow::ipc::RecordBatchStreamWriter}
\title{class arrow::ipc::RecordBatchStreamWriter

Writer for the Arrow streaming binary format}
\description{
class arrow::ipc::RecordBatchStreamWriter

Writer for the Arrow streaming binary format
}
\section{usage}{
\preformatted{writer <- RecordBatchStreamWriter(sink, schema)

writer$write_batch(batch)
writer$write_table(table)
writer$close()
}
}

\section{Factory}{


The \code{\link[=RecordBatchStreamWriter]{RecordBatchStreamWriter()}} function creates a record batch stream writer.
}

\section{Methods}{

inherited from \link[=arrow__ipc__RecordBatchWriter]{arrow::ipc::RecordBatchWriter}
\itemize{
\item \code{$write_batch(batch)}: Write record batch to stream
\item \code{$write_table(table)}: write Table to stream
\item \code{$close()}: close stream
}
}

\keyword{datasets}
