% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\docType{class}
\name{arrow__ipc__RecordBatchWriter}
\alias{arrow__ipc__RecordBatchWriter}
\alias{arrow::ipc::RecordBatchWriter}
\title{class arrow::ipc::RecordBatchWriter}
\description{
class arrow::ipc::RecordBatchWriter
}
\section{Methods}{

\itemize{
\item \code{$write_batch(batch)}: Write record batch to stream
\item \code{$write_table(table)}: write Table to stream
\item \code{$close()}: close stream
}
}

\section{Derived classes}{

\itemize{
\item \link[=arrow__ipc__RecordBatchStreamWriter]{arrow::ipc::RecordBatchStreamWriter} implements the streaming binary format
\item \link[=arrow__ipc__RecordBatchFileWriter]{arrow::ipc::RecordBatchFileWriter} implements the binary file format
}
}

\keyword{datasets}
