% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\docType{class}
\name{arrow__json__TableReader}
\alias{arrow__json__TableReader}
\alias{arrow::json::TableReader}
\title{class arrow::json::TableReader}
\description{
class arrow::json::TableReader
}
\section{Methods}{

\itemize{
\item \code{Read()} : read the JSON file as an \link[=arrow__Table]{arrow::Table}
}
}

\keyword{datasets}
