% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatch.R
\docType{class}
\name{arrow__RecordBatch}
\alias{arrow__RecordBatch}
\alias{arrow::RecordBatch}
\title{class arrow::RecordBatch}
\description{
class arrow::RecordBatch
}
\section{Methods}{


TODO
}

\keyword{datasets}
