% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\docType{class}
\name{arrow__RecordBatchReader}
\alias{arrow__RecordBatchReader}
\alias{arrow::RecordBatchReader}
\title{class arrow::RecordBatchReader}
\description{
class arrow::RecordBatchReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
