% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema.R
\docType{class}
\name{arrow__Schema}
\alias{arrow__Schema}
\alias{arrow::Schema}
\title{class arrow::Schema}
\description{
class arrow::Schema
}
\section{Usage}{
\preformatted{s <- schema(...)

s$ToString()
s$num_fields()
s$field(i)
}
}

\section{Methods}{

\itemize{
\item \code{$ToString()}: convert to a string
\item \code{$num_fields()}: returns the number of fields
\item \code{$field(i)}: returns the field at index \code{i} (0-based)
}
}

\keyword{datasets}
