% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arrow_array}
\alias{arrow_array}
\title{create an \link[=arrow__Array]{arrow::Array} from an R vector}
\usage{
arrow_array(x, type = NULL)
}
\arguments{
\item{x}{R object}

\item{type}{Explicit \link[=arrow__DataType]{type}, or NULL (the default) to infer
from the data}
}
\value{
An Arrow array R6 object
}
\description{
create an \link[=arrow__Array]{arrow::Array} from an R vector
}
