% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{arrow_cancellable}
\alias{arrow_cancellable}
\title{Run Arrow code with the ability to cancel}
\usage{
arrow_cancellable(expr)
}
\arguments{
\item{expr}{An expression to evaluate with cancellation}
}
\value{
The value of \code{expr}
}
\description{
Run Arrow code with the ability to cancel
}
