% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{arrow_enable_cancel_from_interrupt}
\alias{arrow_enable_cancel_from_interrupt}
\title{Enable cancel of long-running Arrow operations}
\usage{
arrow_enable_cancel_from_interrupt(enabled = TRUE)
}
\arguments{
\item{enabled}{Use \code{TRUE} to turn on cancellation of long-running
Arrow operations where this is supported; use \code{FALSE} to disable.
Using \code{TRUE} is currently experimental.}
}
\value{
The previous value of \code{enabled}, invisibly
}
\description{
Enable cancel of long-running Arrow operations
}
