% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{arrow_table}
\alias{arrow_table}
\title{Create an \code{arrow::Table} from a data frame}
\usage{
arrow_table(..., schema = NULL)
}
\arguments{
\item{...}{arrays, chunked arrays, or R vectors}

\item{schema}{a schema. The default (\code{NULL}) infers the schema from the \code{...}}
}
\value{
an \code{arrow::Table}
}
\description{
Create an \code{arrow::Table} from a data frame
}
