% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R, R/chunked-array.R
\name{as_arrow_array}
\alias{as_arrow_array}
\alias{as_arrow_array.Array}
\alias{as_arrow_array.ChunkedArray}
\alias{as_arrow_array.default}
\alias{as_chunked_array.ChunkedArray}
\alias{as_chunked_array.Array}
\alias{as_chunked_array.vctrs_vctr}
\title{Convert an object to an Arrow Array}
\usage{
as_arrow_array(x, ..., type = NULL)

\method{as_arrow_array}{Array}(x, ..., type = NULL)

\method{as_arrow_array}{ChunkedArray}(x, ..., type = NULL)

\method{as_arrow_array}{default}(x, ..., type = NULL, from_constructor = FALSE)

\method{as_chunked_array}{ChunkedArray}(x, ..., type = NULL)

\method{as_chunked_array}{Array}(x, ..., type = NULL)

\method{as_chunked_array}{vctrs_vctr}(x, ..., type = NULL)
}
\arguments{
\item{x}{An object to convert to an Arrow Array}

\item{...}{Passed to S3 methods}

\item{type}{A \link[=data-type]{type} for the final Array. A value of \code{NULL}
will default to the type guessed by \code{\link[=type]{type()}}.}
}
\value{
An \link{Array}.
}
\description{
Whereas \code{Array$create()} constructs an \link{Array} from the built-in data types
for which the Arrow package implements fast converters, \code{as_arrow_array()}
provides a means by which other packages can define conversions to Arrow
objects.
}
\examples{
\dontshow{if (arrow_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
as_arrow_array(1:5)
\dontshow{\}) # examplesIf}
}
