% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{as_data_type}
\alias{as_data_type}
\alias{as_data_type.DataType}
\alias{as_data_type.Field}
\alias{as_data_type.Schema}
\title{Convert an object to an Arrow DataType}
\usage{
as_data_type(x, ...)

\method{as_data_type}{DataType}(x, ...)

\method{as_data_type}{Field}(x, ...)

\method{as_data_type}{Schema}(x, ...)
}
\arguments{
\item{x}{An object to convert to an Arrow \link[=data-type]{DataType}}

\item{...}{Passed to S3 methods.}
}
\value{
A \link[=data-type]{DataType} object.
}
\description{
Convert an object to an Arrow DataType
}
\examples{
\dontshow{if (arrow_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
as_data_type(int32())
\dontshow{\}) # examplesIf}
}
