% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-batch-reader.R
\name{as_record_batch_reader}
\alias{as_record_batch_reader}
\alias{as_record_batch_reader.RecordBatchReader}
\title{Convert an object to an Arrow RecordBatchReader}
\usage{
as_record_batch_reader(x, ...)

\method{as_record_batch_reader}{RecordBatchReader}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \link{RecordBatchReader}}

\item{...}{Passed to S3 methods}
}
\value{
A \link{RecordBatchReader}
}
\description{
Convert an object to an Arrow RecordBatchReader
}
\examples{
\dontshow{if (arrow_available() && arrow_with_dataset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
reader <- as_record_batch_reader(data.frame(col1 = 1, col2 = "two"))
reader$read_next_batch()
\dontshow{\}) # examplesIf}
}
