% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{as_schema}
\alias{as_schema}
\alias{as_schema.Schema}
\alias{as_schema.StructType}
\title{Convert an object to an Arrow Schema}
\usage{
as_schema(x, ...)

\method{as_schema}{Schema}(x, ...)

\method{as_schema}{StructType}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \code{\link[=schema]{schema()}}}

\item{...}{Passed to S3 methods.}
}
\value{
A \link{Schema} object.
}
\description{
Convert an object to an Arrow Schema
}
\examples{
as_schema(schema(col1 = int32()))

}
