% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{az_container}
\alias{az_container}
\title{Connect to an Azure Blob Storage container}
\usage{
az_container(container_path, ...)
}
\arguments{
\item{container_path}{string Container name or path}

\item{...}{Additional connection options, passed to \code{AzureFileSystem$create()}}
}
\value{
A \code{SubTreeFileSystem} containing an \code{AzureFileSystem} and the container's
relative path. Note that this function's success does not guarantee that you
are authorized to access the container's contents.
}
\description{
\code{az_conainer} is a convenience function to create an \code{AzureFileSystem} object
that provides a file system interface for blob storage containers in an Azure
Storage Account.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
container_fs <- az_container(
  container_path = "arrow-datasets",
  account_name = azurite_account_name,
  account_key = azurite_account_key,
  blob_storage_authority = azurite_blob_storage_authority,
  blob_storage_scheme = azurite_blob_storage_scheme
)
\dontshow{\}) # examplesIf}
}
