% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\docType{class}
\name{buffer}
\alias{buffer}
\alias{Buffer}
\title{class Buffer}
\usage{
buffer(x)
}
\arguments{
\item{x}{R object. Only raw, numeric and integer vectors are currently supported}
}
\value{
an instance of \code{Buffer} that borrows memory from \code{x}
}
\description{
\code{buffer()} lets you create an \code{arrow::Buffer} from an R object
}
\section{Methods}{

\itemize{
\item \code{$is_mutable()} :
\item \code{$ZeroPadding()} :
\item \code{$size()} :
\item \code{$capacity()}:
}
}

\keyword{datasets}
