% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{buffer_reader}
\alias{buffer_reader}
\title{Create a \code{arrow::BufferReader}}
\usage{
buffer_reader(x)
}
\arguments{
\item{x}{R object to treat as a buffer or a buffer created by \code{\link[=buffer]{buffer()}}}
}
\description{
Create a \code{arrow::BufferReader}
}
