% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{BufferOutputStream}
\alias{BufferOutputStream}
\title{class arrow::io::BufferOutputStream}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\description{
class arrow::io::BufferOutputStream
}
\section{Methods}{


TODO
}

\keyword{datasets}
