% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{BufferReader}
\alias{BufferReader}
\title{Create a \link[=arrow__io__BufferReader]{arrow::io::BufferReader}}
\usage{
BufferReader(x)
}
\arguments{
\item{x}{R object to treat as a buffer or a buffer created by \code{\link[=buffer]{buffer()}}}
}
\description{
Create a \link[=arrow__io__BufferReader]{arrow::io::BufferReader}
}
