% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{cast_options}
\alias{cast_options}
\title{Cast options}
\usage{
cast_options(safe = TRUE, ...)
}
\arguments{
\item{safe}{logical: enforce safe conversion? Default \code{TRUE}}

\item{...}{additional cast options, such as \code{allow_int_overflow},
\code{allow_time_truncate}, and \code{allow_float_truncate}, which are set to \code{!safe}
by default}
}
\value{
A list
}
\description{
Cast options
}
\keyword{internal}
