% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkedArray.R
\name{chunked_array}
\alias{chunked_array}
\title{Create a \link[=chunked-array]{ChunkedArray} from various R vectors}
\usage{
chunked_array(..., type = NULL)
}
\arguments{
\item{\dots}{Vectors to coerce}

\item{type}{currently ignored}
}
\description{
Create a \link[=chunked-array]{ChunkedArray} from various R vectors
}
