% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\docType{class}
\name{Codec}
\alias{Codec}
\title{Compression Codec class}
\description{
Codecs allow you to create \link[=compression]{compressed input and output streams}.
}
\section{Factory}{

The \code{Codec$create()} factory method takes the following argument:
\itemize{
\item \code{type}: string name of the compression method. See \link{CompressionType} for
a list of possible values. \code{type} may be upper- or lower-cased. Support
for compression methods depends on build-time flags for the C++ library.
Most builds support at least "gzip" and "snappy".
\item \code{compression_level}: compression level, the default value (\code{NA}) uses the default
compression level for the selected compression \code{type}.
}
}

