% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{codec_is_available}
\alias{codec_is_available}
\title{Check whether a compression codec is available}
\usage{
codec_is_available(type)
}
\arguments{
\item{type}{A string, one of "uncompressed", "snappy", "gzip", "brotli",
"zstd", "lz4", "lzo", or "bz2", case insensitive.}
}
\value{
Logical: is \code{type} available?
}
\description{
Support for compression libraries depends on the build-time settings of
the Arrow C++ library. This function lets you know which are available for
use.
}
\examples{
\dontshow{if (arrow_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
codec_is_available("gzip")
\dontshow{\}) # examplesIf}
}
