% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compressed_input_stream}
\alias{compressed_input_stream}
\title{Compressed input stream}
\usage{
compressed_input_stream(stream, codec = codec("GZIP"))
}
\arguments{
\item{stream}{Underlying raw input stream}

\item{codec}{a codec}
}
\description{
Compressed input stream
}
