% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{CompressedInputStream}
\alias{CompressedInputStream}
\title{Compressed input stream}
\usage{
CompressedInputStream(stream, codec = codec("GZIP"))
}
\arguments{
\item{stream}{Underlying raw input stream}

\item{codec}{a codec}
}
\description{
Compressed input stream
}
