% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{CompressedOutputStream}
\alias{CompressedOutputStream}
\title{Compressed output stream}
\usage{
CompressedOutputStream(stream, codec = compression_codec("GZIP"))
}
\arguments{
\item{stream}{Underlying raw output stream}

\item{codec}{a codec}
}
\description{
Compressed output stream
}
\details{
This function is not supported in Windows.
}
