% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\docType{class}
\name{compression}
\alias{compression}
\alias{CompressedOutputStream}
\alias{CompressedInputStream}
\title{Compressed stream classes}
\description{
\code{CompressedInputStream} and \code{CompressedOutputStream}
allow you to apply a compression \link{Codec} to an
input or output stream.
}
\section{Factory}{


The \code{CompressedInputStream$create()} and \code{CompressedOutputStream$create()}
factory methods instantiate the object and take the following arguments:
\itemize{
\item \code{stream} An \link{InputStream} or \link{OutputStream}, respectively
\item \code{codec} A \code{Codec}, either a \link{Codec} instance or a string
\item \code{compression_level} compression level for when the \code{codec} argument is given as a string
}
}

\section{Methods}{


Methods are inherited from \link{InputStream} and \link{OutputStream}, respectively
}

