% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-funcs-string.R
\name{contains_regex}
\alias{contains_regex}
\title{Does this string contain regex metacharacters?}
\usage{
contains_regex(string)
}
\arguments{
\item{string}{String to be tested}
}
\value{
Logical: does \code{string} contain regex metacharacters?
}
\description{
Does this string contain regex metacharacters?
}
\keyword{internal}
