% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cpu_count}
\alias{cpu_count}
\title{Return number of threads used by the global CPU thread pool in libarrow}
\usage{
cpu_count()
}
\description{
Return number of threads used by the global CPU thread pool in libarrow
}
