% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R, R/json.R
\docType{class}
\name{CsvTableReader}
\alias{CsvTableReader}
\alias{JsonTableReader}
\title{Arrow CSV and JSON table reader classes}
\description{
\code{CsvTableReader} and \code{JsonTableReader} wrap the Arrow C++ CSV
and JSON table readers. See their usage in \code{\link[=read_csv_arrow]{read_csv_arrow()}} and
\code{\link[=read_json_arrow]{read_json_arrow()}}, respectively.
}
\section{Factory}{


The \code{CsvTableReader$create()} and \code{JsonTableReader$create()} factory methods
take the following arguments:
\itemize{
\item \code{file} An Arrow \link{InputStream}
\item \code{convert_options} (CSV only), \code{parse_options}, \code{read_options}: see
\link{CsvReadOptions}
\item \code{...} additional parameters.
}
}

\section{Methods}{

\itemize{
\item \verb{$Read()}: returns an Arrow Table.
}
}

