% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{DataSource}
\alias{DataSource}
\alias{DataSourceDiscovery}
\alias{FileSystemDataSourceDiscovery}
\title{Data sources for a Dataset}
\description{
A \link{Dataset} can have one or more \code{DataSource}s. A \code{DataSource} contains one
or more \code{DataFragments}, such as files, of a common type and partition
scheme. \code{DataSourceDiscovery} is used to create a \code{DataSource}, inspect the
\link{Schema} of the fragments contained in it, and declare a partition scheme.
\code{FileSystemDataSourceDiscovery} is a subclass of \code{DataSourceDiscovery} for
discovering files in the local file system, the only currently supported
file system.

In general, you'll deal with \code{DataSourceDiscovery} rather than \code{DataSource}
itself.

Return the DataSource's \code{Schema}
}
\section{Factory}{

For the \code{DataSourceDiscovery$create()} factory method, see \code{\link[=data_source]{data_source()}},
an alias for it.

\code{FileSystemDataSourceDiscovery$create()} is a lower-level factory method and
takes the following arguments:
\itemize{
\item \code{filesystem}: A \link{FileSystem}
\item \code{selector}: A \link{FileSelector}
\item \code{format}: Currently only "parquet" is supported
}
}

\section{Methods}{

\code{DataSource} has one defined method:
\itemize{
\item \verb{$schema}: Active binding, returns the \link{Schema} of the DataSource
}

\code{DataSourceDiscovery} and its subclasses have the following methods:
\itemize{
\item \verb{$Inspect()}: Walks the files in the directory and returns a common \link{Schema}
\item \verb{$Finish(schema)}: Returns a \code{DataSource}
}
}

\seealso{
\link{Dataset} for what do do with a \code{DataSource}
}
