% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\docType{class}
\name{DataType}
\alias{DataType}
\title{DataType class}
\description{
DataType class
}
\section{R6 Methods}{

\itemize{
\item \verb{$ToString()}: String representation of the DataType
\item \verb{$Equals(other)}: Is the DataType equal to \code{other}
\item \verb{$fields()}: The children fields associated with this type
\item \verb{$code(namespace)}: Produces an R call of the data type. Use \code{namespace=TRUE} to call with \verb{arrow::}.
}

There are also some active bindings:
\itemize{
\item \verb{$id}: integer Arrow type id.
\item \verb{$name}: string Arrow type name.
\item \verb{$num_fields}: number of child fields.
}
}

\seealso{
\code{\link[=infer_type]{infer_type()}}

\code{\link{data-type}}
}
