% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{int8}
\alias{int8}
\alias{int16}
\alias{int32}
\alias{int64}
\alias{uint8}
\alias{uint16}
\alias{uint32}
\alias{uint64}
\alias{float16}
\alias{float32}
\alias{float64}
\alias{boolean}
\alias{utf8}
\alias{date32}
\alias{date64}
\alias{time32}
\alias{time64}
\alias{null}
\alias{timestamp}
\alias{struct}
\alias{schema}
\alias{list_of}
\title{Apache Arrow data types}
\usage{
int8()

int16()

int32()

int64()

uint8()

uint16()

uint32()

uint64()

float16()

float32()

float64()

boolean()

utf8()

date32()

date64()

time32(unit)

time64(unit)

null()

timestamp(unit, timezone)

struct(...)

schema(...)

list_of(type)
}
\arguments{
\item{unit}{time unit}

\item{...}{...}

\item{type}{type}

\item{precision}{precision}

\item{scale}{scale}
}
\description{
Apache Arrow data types
}
