% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory-pool.R
\name{default_memory_pool}
\alias{default_memory_pool}
\title{Arrow's default \link{MemoryPool}}
\usage{
default_memory_pool()
}
\value{
the default \link{MemoryPool}
}
\description{
Arrow's default \link{MemoryPool}
}
\keyword{internal}
