% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime-info.R
\name{default_runtime_info}
\alias{default_runtime_info}
\title{Arrow's default \link{RuntimeInfo}}
\usage{
default_runtime_info()
}
\value{
the default \link{RuntimeInfo}
}
\description{
Arrow's default \link{RuntimeInfo}
}
\keyword{internal}
