% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Create a dictionary type}
\usage{
dictionary(index_type = int32(), value_type = utf8(), ordered = FALSE)
}
\arguments{
\item{index_type}{A DataType for the indices (default \code{\link[=int32]{int32()}})}

\item{value_type}{A DataType for the values (default \code{\link[=utf8]{utf8()}})}

\item{ordered}{Is this an ordered dictionary (default \code{FALSE})?}
}
\value{
A \link{DictionaryType}
}
\description{
Create a dictionary type
}
\seealso{
\link[=data-type]{Other Arrow data types}
}
