% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{enums}
\alias{enums}
\alias{TimeUnit}
\alias{DateUnit}
\alias{Type}
\alias{StatusCode}
\alias{FileMode}
\alias{MessageType}
\alias{CompressionType}
\alias{FileType}
\alias{ParquetVersionType}
\alias{MetadataVersion}
\alias{QuantileInterpolation}
\alias{NullEncodingBehavior}
\alias{NullHandlingBehavior}
\alias{RoundMode}
\alias{JoinType}
\title{Arrow enums}
\usage{
TimeUnit

DateUnit

Type

StatusCode

FileMode

MessageType

CompressionType

FileType

ParquetVersionType

MetadataVersion

QuantileInterpolation

NullEncodingBehavior

NullHandlingBehavior

RoundMode

JoinType
}
\description{
Arrow enums
}
\keyword{internal}
