% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\docType{class}
\name{ExtensionArray}
\alias{ExtensionArray}
\title{class arrow::ExtensionArray}
\description{
class arrow::ExtensionArray
}
\section{Methods}{


The \code{ExtensionArray} class inherits from \code{Array}, but also provides
access to the underlying storage of the extension.
\itemize{
\item \verb{$storage()}: Returns the underlying \link{Array} used to store
values.
}

The \code{ExtensionArray} is not intended to be subclassed for extension
types.
}

