% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{extract_schema_code}
\alias{extract_schema_code}
\title{Extract code required to specify a schema}
\usage{
extract_schema_code(schema, namespace = FALSE)
}
\arguments{
\item{schema}{An Arrow Schema}

\item{namespace}{Prefix type function calls with the arrow namespace?}
}
\description{
This can be useful when creating a custom schema using the inferred schema as a starting point.
}
\examples{
my_schema <- schema(x = int8(), y = string())
extract_schema_code(my_schema)
}
