% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{feather_table_reader}
\alias{feather_table_reader}
\title{A arrow::ipc::feather::TableReader to read from a file}
\usage{
feather_table_reader(file, mmap = TRUE, ...)
}
\arguments{
\item{file}{A file path, arrow::io::RandomAccessFile}

\item{mmap}{Is the file memory mapped (applicable to the character and fs_path methods)}

\item{...}{extra parameters}
}
\description{
A arrow::ipc::feather::TableReader to read from a file
}
