% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{feather_table_writer}
\alias{feather_table_writer}
\title{Create TableWriter that writes into a stream}
\usage{
feather_table_writer(stream)
}
\arguments{
\item{stream}{an OutputStream}
}
\description{
Create TableWriter that writes into a stream
}
