% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\docType{class}
\name{FeatherReader}
\alias{FeatherReader}
\title{FeatherReader class}
\description{
This class enables you to interact with Feather files. Create
one to connect to a file or other InputStream, and call \code{Read()} on it to
make an \code{arrow::Table}. See its usage in \code{\link[=read_feather]{read_feather()}}.
}
\section{Factory}{


The \code{FeatherReader$create()} factory method instantiates the object and
takes the following argument:
\itemize{
\item \code{file} an Arrow file connection object inheriting from \code{RandomAccessFile}.
}
}

\section{Methods}{

\itemize{
\item \verb{$Read(columns)}: Returns a \code{Table} of the selected columns, a vector of
integer indices
\item \verb{$column_names}: Active binding, returns the column names in the Feather file
\item \verb{$schema}: Active binding, returns the schema of the Feather file
\item \verb{$version}: Active binding, returns \code{1} or \code{2}, according to the Feather
file version
}
}

