% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{field}
\alias{field}
\title{Create a Field}
\usage{
field(name, type, metadata = NULL, nullable = TRUE)
}
\arguments{
\item{name}{field name}

\item{type}{logical type, instance of \link{DataType}}

\item{metadata}{a named character vector or list to attach as field metadata.
All values will be coerced to \code{character}.}

\item{nullable}{TRUE if field is nullable}
}
\description{
Create a Field
}
\examples{
field("x", int32())
field("x", int32(), metadata = list(key = "value"))
}
\seealso{
\link{Field}
}
