% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{FileOutputStream}
\alias{FileOutputStream}
\title{Open a \link[=arrow__io__FileOutputStream]{arrow::io::FileOutputStream}}
\usage{
FileOutputStream(path)
}
\arguments{
\item{path}{file path}
}
\value{
a \link[=arrow__io__FileOutputStream]{arrow::io::FileOutputStream}
}
\description{
Open a \link[=arrow__io__FileOutputStream]{arrow::io::FileOutputStream}
}
