% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{FileSelector}
\alias{FileSelector}
\title{file selector}
\description{
file selector
}
\section{Factory}{


The \verb{$create()} factory method instantiates a \code{FileSelector} given the 3 fields
described below.
}

\section{Fields}{

\itemize{
\item \code{base_dir}: The directory in which to select files. If the path exists but
doesn't point to a directory, this should be an error.
\item \code{allow_not_found}: The behavior if \code{base_dir} doesn't exist in the
filesystem. If \code{FALSE}, an error is returned.  If \code{TRUE}, an empty
selection is returned
\item \code{recursive}: Whether to recurse into subdirectories.
}
}

