% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\docType{data}
\name{FileStats}
\alias{FileStats}
\title{FileSystem entry stats}
\description{
FileSystem entry stats
}
\section{Methods}{

\itemize{
\item \code{base_name()} : The file base name (component after the last directory
separator).
\item \code{extension()} : The file extension
}
}

\section{Active bindings}{

\itemize{
\item \code{$type}: The file type
\item \code{$path}: The full file path in the filesystem
\item \code{$size}: The size in bytes, if available.  Only regular files are
guaranteed to have a size.
\item \code{$mtime}: The time of last modification, if available.
}
}

\keyword{datasets}
