% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-write.R
\name{FileWriteOptions}
\alias{FileWriteOptions}
\alias{ParquetFileWriteOptions}
\title{Format-specific write options}
\description{
A \code{FileWriteOptions} holds write options specific to a \code{FileFormat}.
}
