% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\docType{class}
\name{FixedWidthType}
\alias{FixedWidthType}
\title{FixedWidthType class}
\description{
\code{FixedWidthType} is a base class for data types with a fixed width in bits.
This includes all integer types, floating-point types, \code{Boolean},
\code{FixedSizeBinary}, temporal types (dates, times, timestamps, durations),
and decimal types.
}
\section{R6 Methods}{


\code{FixedWidthType} inherits from \link{DataType}, so it has the same methods.
}

\section{Active bindings}{

\itemize{
\item \verb{$bit_width}: The width of the type in bits
}
}

\keyword{internal}
