% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_connect}
\alias{flight_connect}
\title{Connect to a Flight server}
\usage{
flight_connect(host = "localhost", port, scheme = "grpc+tcp")
}
\arguments{
\item{host}{string hostname to connect to}

\item{port}{integer port to connect on}

\item{scheme}{URL scheme, default is "grpc+tcp"}
}
\value{
A \code{pyarrow.flight.FlightClient}.
}
\description{
Connect to a Flight server
}
