% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_disconnect}
\alias{flight_disconnect}
\title{Explicitly close a Flight client}
\usage{
flight_disconnect(client)
}
\arguments{
\item{client}{The client to disconnect}
}
\description{
Explicitly close a Flight client
}
