% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{format_schema}
\alias{format_schema}
\title{Get a string representing a Dataset or RecordBatchReader object's schema}
\usage{
format_schema(obj)
}
\arguments{
\item{obj}{a Dataset or RecordBatchReader}
}
\value{
A string containing a formatted representation of the schema of \code{obj}
}
\description{
Get a string representing a Dataset or RecordBatchReader object's schema
}
\keyword{internal}
