% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-format.R
\name{FragmentScanOptions}
\alias{FragmentScanOptions}
\alias{CsvFragmentScanOptions}
\title{Format-specific scan options}
\description{
A \code{FragmentScanOptions} holds options specific to a \code{FileFormat} and a scan
operation.
}
\section{Factory}{

\code{FragmentScanOptions$create()} takes the following arguments:
\itemize{
\item \code{format}: A string identifier of the file format. Currently supported values:
\itemize{
\item "csv"/"text", aliases for the same format.
}
\item \code{...}: Additional format-specific options

\code{format = "text"}: see \link{CsvConvertOptions}. Note that options can only be
specified with the Arrow C++ library naming. Also, "block_size" from
\link{CsvReadOptions} may be given.
}

It returns the appropriate subclass of \code{FragmentScanOptions}
(e.g. \code{CsvFragmentScanOptions}).
}

