% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_copy_file}
\alias{fs_copy_file}
\title{Copy a file in a file system}
\usage{
fs_copy_file(fs, src, dest)
}
\arguments{
\item{fs}{A FileSystem object}

\item{src}{Source path}

\item{dest}{Destination path}
}
\description{
Copy a file in a file system
}
