% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_create_dir}
\alias{fs_create_dir}
\title{Create a file in a file system}
\usage{
fs_create_dir(fs, path, recursive = TRUE)
}
\arguments{
\item{fs}{A FileSystem object}

\item{path}{A character vector of one or more paths}

\item{recursive}{Create intermediate directories if they do not exist}
}
\description{
Create a file in a file system
}
