% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_delete_dir}
\alias{fs_delete_dir}
\title{Delete a file in a file system}
\usage{
fs_delete_dir(fs, path)
}
\arguments{
\item{fs}{A FileSystem object}

\item{path}{A character vector of one or more paths}
}
\description{
Delete a file in a file system
}
