% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_get_file_info}
\alias{fs_get_file_info}
\title{Get file metadata}
\usage{
fs_get_file_info(fs, path)
}
\arguments{
\item{fs}{A FileSystem object}

\item{path}{A character vector of one or more paths}
}
\description{
Get file metadata
}
