% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_ls}
\alias{fs_ls}
\title{List files in a file system}
\usage{
fs_ls(fs, path = ".", ...)
}
\arguments{
\item{fs}{A FileSystem object}

\item{path}{A character vector of one or more paths}

\item{...}{Additional arguments passed to \code{\link[=FileSelector$create]{FileSelector$create()}}}
}
\description{
List files in a file system
}
