% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_make_path}
\alias{fs_make_path}
\title{Construct path to file system path or directory}
\usage{
fs_make_path(fs, path)
}
\arguments{
\item{fs}{A FileSystem object}

\item{path}{A character vector of one or more paths}
}
\description{
Create a \code{SubTreeFileSystem} from the current \code{FileSystem} rooted at the specified path.
}
