% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{fs_move_file}
\alias{fs_move_file}
\title{Move or rename a file in a file system}
\usage{
fs_move_file(fs, src, dest)
}
\arguments{
\item{fs}{A FileSystem object}

\item{src}{Source path}

\item{dest}{Destination path}
}
\description{
Move / rename a file or directory. If the destination exists:
\itemize{
\item if it is a non-empty directory, an error is returned
\item otherwise, if it has the same type as the source, it is replaced
\item otherwise, behavior is unspecified (implementation-dependent).
}
}
