% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pkg_ns}
\alias{get_pkg_ns}
\title{Prepares the value of \code{.ns} to use with \code{rlang::call2()}}
\usage{
get_pkg_ns(explicit_pkg_name)
}
\arguments{
\item{explicit_pkg_name}{Whether to include the package name or not.}
}
\description{
Prepares the value of \code{.ns} to use with \code{rlang::call2()}
}
