% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-funcs-string.R
\name{get_stringr_pattern_options}
\alias{get_stringr_pattern_options}
\title{Get \code{stringr} pattern options}
\usage{
get_stringr_pattern_options(pattern)
}
\arguments{
\item{pattern}{Unevaluated expression containing a call to a \code{stringr}
pattern modifier function}
}
\value{
List containing elements \code{pattern}, \code{fixed}, and \code{ignore_case}
}
\description{
This function assigns definitions for the \code{stringr} pattern modifier
functions (\code{fixed()}, \code{regex()}, etc.) inside itself, and uses them to
evaluate the quoted expression \code{pattern}, returning a list that is used
to control pattern matching behavior in internal \code{arrow} functions.
}
\keyword{internal}
