% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetCpuThreadPoolCapacity}
\alias{GetCpuThreadPoolCapacity}
\title{Get the capacity of the global thread pool}
\usage{
GetCpuThreadPoolCapacity()
}
\value{
the number of worker threads in the thread pool to which
Arrow dispatches various CPU-bound tasks. This is an ideal number,
not necessarily the exact number of threads at a given point in time.

You can change this number using \code{\link[=SetCpuThreadPoolCapacity]{SetCpuThreadPoolCapacity()}}.
}
\description{
Get the capacity of the global thread pool
}
