% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{gs_bucket}
\alias{gs_bucket}
\title{Connect to a Google Cloud Storage (GCS) bucket}
\usage{
gs_bucket(bucket, ...)
}
\arguments{
\item{bucket}{string GCS bucket name or path}

\item{...}{Additional connection options, passed to \code{GcsFileSystem$create()}}
}
\value{
A \code{SubTreeFileSystem} containing an \code{GcsFileSystem} and the bucket's
relative path. Note that this function's success does not guarantee that you
are authorized to access the bucket's contents.
}
\description{
\code{gs_bucket()} is a convenience function to create an \code{GcsFileSystem} object
that holds onto its relative path
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bucket <- gs_bucket("voltrondata-labs-datasets")
\dontshow{\}) # examplesIf}
}
