% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{infer_schema}
\alias{infer_schema}
\title{Extract a schema from an object}
\usage{
infer_schema(x)
}
\arguments{
\item{x}{An object which has a schema, e.g. a \code{Dataset}}
}
\description{
Extract a schema from an object
}
