% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{install_pyarrow}
\alias{install_pyarrow}
\title{Install pyarrow for use with reticulate}
\usage{
install_pyarrow(envname = NULL, nightly = FALSE, ...)
}
\arguments{
\item{envname}{The name or full path of the Python environment to install
into. This can be a virtualenv or conda environment created by \code{reticulate}.
See \code{reticulate::py_install()}.}

\item{nightly}{logical: Should we install a development version of the
package? Default is to use the official release version.}

\item{...}{additional arguments passed to \code{reticulate::py_install()}.}
}
\description{
\code{pyarrow} is the Python package for Apache Arrow. This function helps with
installing it for use with \code{reticulate}.
}
