% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{io_thread_count}
\alias{io_thread_count}
\alias{set_io_thread_count}
\title{Manage the global I/O thread pool in libarrow}
\usage{
io_thread_count()

set_io_thread_count(num_threads)
}
\arguments{
\item{num_threads}{integer: New number of threads for thread pool. At least
two threads are recommended to support all operations in the arrow
package.}
}
\description{
Manage the global I/O thread pool in libarrow
}
