% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{is_in}
\alias{is_in}
\title{\code{is_in} for Arrow objects}
\usage{
is_in(x, table, ...)
}
\arguments{
\item{x}{\code{Array} or \code{ChunkedArray}}

\item{table}{\code{Array}, \code{ChunkedArray}, or R vector lookup table.}

\item{...}{additional arguments, ignored}
}
\value{
An \code{boolean}-type \code{Array} of the same length as \code{x} with
values indicating per element of \code{x} it it is present in \code{table}.
}
\description{
\code{\%in\%} is not a generic so, so we can't just define Arrow methods for
it. This function exposes the analogous function in the Arrow C++ library.
}
