% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json_parse_options}
\alias{json_parse_options}
\title{Parsing options}
\usage{
json_parse_options(newlines_in_values = FALSE)
}
\arguments{
\item{newlines_in_values}{Whether objects may be printed across multiple lines (for example pretty printed),}
}
\description{
Parsing options
}
