% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json_read_options}
\alias{json_read_options}
\title{read options for the json reader}
\usage{
json_read_options(use_threads = TRUE, block_size = 1048576L)
}
\arguments{
\item{use_threads}{Whether to use the global CPU thread pool}

\item{block_size}{Block size we request from the IO layer; also determines the size of chunks when use_threads is \code{TRUE}. NB: if false, input must end with an empty line}
}
\description{
read options for the json reader
}
