% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json_table_reader}
\alias{json_table_reader}
\title{Json table reader}
\usage{
json_table_reader(file, read_options = json_read_options(),
  parse_options = json_parse_options(), ...)
}
\arguments{
\item{file}{file}

\item{read_options, }{see \code{\link[=json_read_options]{json_read_options()}}}

\item{parse_options, }{see \code{\link[=json_parse_options]{json_parse_options()}}}

\item{...}{additional parameters.}
}
\description{
Json table reader
}
