% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-format.R
\name{JsonFileFormat}
\alias{JsonFileFormat}
\title{JSON dataset file format}
\value{
A \code{JsonFileFormat} object
}
\description{
A \code{JsonFileFormat} is a \link{FileFormat} subclass which holds information about how to
read and parse the files included in a JSON \code{Dataset}.
}
\section{Factory}{

\code{JsonFileFormat$create()} can take options in the form of lists passed through as \code{parse_options},
or \code{read_options} parameters.
}

\examples{
\dontshow{if (arrow_with_dataset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
\seealso{
\link{FileFormat}
}
