% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\docType{class}
\name{JsonTableReader}
\alias{JsonTableReader}
\title{class JsonTableReader}
\description{
class JsonTableReader
}
\section{Methods}{

\itemize{
\item \code{Read()} : read the JSON file as an \link[=Table]{arrow::Table}
}
}

\keyword{datasets}
