% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{list_compute_functions}
\alias{list_compute_functions}
\title{List available Arrow C++ compute functions}
\usage{
list_compute_functions(pattern = NULL, ...)
}
\arguments{
\item{pattern}{Optional regular expression to filter the function list}

\item{...}{Additional parameters passed to \code{grep()}}
}
\value{
A character vector of available Arrow C++ function names
}
\description{
This function lists the names of all available Arrow compute functions.
These can be called by passing to \code{\link[=call_function]{call_function()}}, or they can be
called by name with an \code{arrow_} prefix inside a \code{dplyr} verb.
}
\details{
The resulting list describes the capabilities of your \code{arrow} build.
Some functions, such as string and regular expression functions,
require optional build-time C++ dependencies. If your \code{arrow} package
was not compiled with those features enabled, those functions will
not appear in this list.
}
