% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{list_flights}
\alias{list_flights}
\alias{flight_path_exists}
\title{See available resources on a Flight server}
\usage{
list_flights(client)

flight_path_exists(client, path)
}
\arguments{
\item{client}{\code{pyarrow.flight.FlightClient}, as returned by \code{\link[=flight_connect]{flight_connect()}}}

\item{path}{string identifier under which data is stored}
}
\value{
\code{list_flights()} returns a character vector of paths.
\code{flight_path_exists()} returns a logical value, the equivalent of \code{path \%in\% list_flights()}
}
\description{
See available resources on a Flight server
}
