% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{load_flight_server}
\alias{load_flight_server}
\title{Load a Python Flight server}
\usage{
load_flight_server(name, path = system.file(package = "arrow"))
}
\arguments{
\item{name}{string Python module name}

\item{path}{file system path where the Python module is found. Default is
to look in the \verb{inst/} directory for included modules.}
}
\description{
Load a Python Flight server
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
load_flight_server("demo_flight_server")
\dontshow{\}) # examplesIf}
}
