% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\docType{class}
\name{LocalFileSystem}
\alias{LocalFileSystem}
\title{LocalFileSystem class}
\description{
EXPERIMENTAL: a \link{FileSystem} implementation accessing files
on the local machine.
}
\keyword{datasets}
