% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message_reader}
\alias{message_reader}
\title{Open a MessageReader that reads from a stream}
\usage{
message_reader(stream)
}
\arguments{
\item{stream}{an InputStream}
}
\description{
Open a MessageReader that reads from a stream
}
