% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\docType{class}
\name{MessageReader}
\alias{MessageReader}
\title{MessageReader class}
\description{
\code{MessageReader} reads \code{Message} objects from an input stream.
}
\section{R6 Methods}{

\itemize{
\item \verb{$ReadNextMessage()}: Read the next \code{Message} from the stream. Returns \code{NULL} if
there are no more messages.
}
}

\section{Factory}{


\code{MessageReader$create()} takes the following argument:
\itemize{
\item \code{stream}: An \link{InputStream} or object coercible to one (e.g., a raw vector)
}
}

