% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mmap_create}
\alias{mmap_create}
\title{Create a new read/write memory mapped file of a given size}
\usage{
mmap_create(path, size)
}
\arguments{
\item{path}{file path}

\item{size}{size in bytes}
}
\value{
a \link[=arrow__io__MemoryMappedFile]{arrow::io::MemoryMappedFile}
}
\description{
Create a new read/write memory mapped file of a given size
}
