% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mmap_open}
\alias{mmap_open}
\title{Open a memory mapped file}
\usage{
mmap_open(path, mode = c("read", "write", "readwrite"))
}
\arguments{
\item{path}{file path}

\item{mode}{file mode (read/write/readwrite)}
}
\description{
Open a memory mapped file
}
