% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{MockOutputStream}
\alias{MockOutputStream}
\title{Open a \link[=arrow__io__MockOutputStream]{arrow::io::MockOutputStream}}
\usage{
MockOutputStream()
}
\value{
a \link[=arrow__io__MockOutputStream]{arrow::io::MockOutputStream}
}
\description{
Open a \link[=arrow__io__MockOutputStream]{arrow::io::MockOutputStream}
}
