% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{parquet_arrow_reader_properties}
\alias{parquet_arrow_reader_properties}
\title{Create a new ArrowReaderProperties instance}
\usage{
parquet_arrow_reader_properties(use_threads = option_use_threads())
}
\arguments{
\item{use_threads}{use threads?}
}
\description{
Create a new ArrowReaderProperties instance
}
\keyword{internal}
