% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{ParquetReaderProperties}
\alias{ParquetReaderProperties}
\title{ParquetReaderProperties class
TODO: add in relevant methods etc using the ParquetArrowReaderProperties docs}
\description{
ParquetReaderProperties class
TODO: add in relevant methods etc using the ParquetArrowReaderProperties docs
}
