% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_compact_col_spec}
\alias{parse_compact_col_spec}
\title{Parse a compact column type specification into Arrow schema}
\usage{
parse_compact_col_spec(col_types, col_names)
}
\arguments{
\item{col_types}{A compact string representation of the column types,
an Arrow \link{Schema}, or \code{NULL} (the default) to infer types from the data.}

\item{col_names}{If \code{TRUE}, the first row of the input will be used as the
column names and will not be included in the data frame. If \code{FALSE}, column
names will be generated by Arrow, starting with "f0", "f1", ..., "fN".
Alternatively, you can specify a character vector of column names.}
}
\value{
A Schema object
}
\description{
Parse a compact column type specification into Arrow schema
}
\examples{
parse_compact_col_spec(col_types = "ci", colnames = c("x", "y"))

}
\keyword{internal}
