% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Partitioning}
\alias{Partitioning}
\alias{DirectoryPartitioning}
\alias{HivePartitioning}
\title{Define Partitioning for a Source}
\description{
Pass a \code{Partitioning} object to a \link{FileSystemSourceFactory}'s \verb{$create()}
method to indicate how the file's paths should be interpreted to define
partitioning.

\code{DirectoryPartitioning} describes how to interpret raw path segments, in
order. For example, \code{schema(year = int16(), month = int8())} would define
partitions for file paths like "2019/01/file.parquet",
"2019/02/file.parquet", etc.

\code{HivePartitioning} is for Hive-style partitioning, which embeds field
names and values in path segments, such as
"/year=2019/month=2/data.parquet". Because fields are named in the path
segments, order does not matter.
}
\section{Factory}{

Both \code{DirectoryPartitioning$create()} and \code{HivePartitioning$create()}
factory methods take a \link{Schema} as a single input argument. The helper
function \code{hive_partition(...)} is shorthand for
\code{HivePartitioning$create(schema(...))}.
}

