% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{push_data}
\alias{push_data}
\title{Send data to a Flight server}
\usage{
push_data(client, data, path)
}
\arguments{
\item{client}{\code{pyarrow.flight.FlightClient}, as returned by \code{\link[=flight_connect]{flight_connect()}}}

\item{data}{\code{data.frame} or \link{RecordBatch} to upload}

\item{path}{string identifier to store the data under}
}
\value{
\code{client}, invisibly.
}
\description{
Send data to a Flight server
}
