% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{r_to_py}
\alias{r_to_py}
\title{Transfer an Arrow memory structure to Python}
\usage{
r_to_py(x, ...)
}
\arguments{
\item{x}{the R-based Arrow object to export to Python}

\item{...}{arguments passed to the methods}
}
\description{
Transfer an Arrow memory structure to Python
}
\keyword{internal}
