% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{read_feather}
\alias{read_feather}
\title{Read a Feather file (deprecated)}
\usage{
read_feather(file, col_select = NULL, as_data_frame = TRUE, mmap = TRUE)
}
\arguments{
\item{file}{A character file name or URI, connection, \code{raw} vector, an
Arrow input stream, or a \code{FileSystem} with path (\code{SubTreeFileSystem}).
If a file name or URI, an Arrow \link{InputStream} will be opened and
closed when finished. If an input stream is provided, it will be left
open.}

\item{col_select}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}

\item{as_data_frame}{Should the function return a \code{tibble} (default) or
an Arrow \link{Table}?}

\item{mmap}{Logical: whether to memory-map the file (default \code{TRUE})}
}
\value{
A \code{tibble} if \code{as_data_frame} is \code{TRUE} (the default), or an
Arrow \link{Table} otherwise
}
\description{
\code{read_feather()} is deprecated and will be removed in a future release.
Use \code{\link[=read_ipc_file]{read_ipc_file()}} instead.

\code{read_feather()} can read both the Feather V1 format (a legacy format which
is also being deprecated) and the Feather V2 format (which is the Arrow IPC format).
\code{read_ipc_file()} can also read both formats.
}
\seealso{
\code{\link[=read_ipc_file]{read_ipc_file()}}
}
