% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{read_json_arrow}
\alias{read_json_arrow}
\title{Read json file into an arrow::Table}
\usage{
read_json_arrow(...)
}
\arguments{
\item{...}{Used to construct an arrow::json::TableReader}
}
\description{
Use arrow::json::TableReader from \code{\link[=json_table_reader]{json_table_reader()}}
}
