% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatch.R
\name{read_record_batch}
\alias{read_record_batch}
\title{Read Record Batch File}
\usage{
read_record_batch(path)
}
\arguments{
\item{stream}{a \code{raw()} containing a stream of record batches.}
}
\description{
Reads an \code{arrow::RecordBatch} from a file using
\code{arrow::RecordBatchFileReader}.
}
