% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatch.R
\name{read_record_batch_stream}
\alias{read_record_batch_stream}
\title{Read Record Batch Stream}
\usage{
read_record_batch_stream(stream)
}
\arguments{
\item{stream}{a \code{raw()} containing a stream of record batches.}
}
\description{
Reads a stream of \code{arrow::RecordBatch} from a \code{raw()} vector
using \code{arrow::RecordBatchStreamReader} and returns a data frame
for each record batch.
}
