% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{read_schema}
\alias{read_schema}
\title{read a Schema from a stream}
\usage{
read_schema(stream, ...)
}
\arguments{
\item{stream}{a \code{Message}, \code{InputStream}, or \code{Buffer}}

\item{...}{currently ignored}
}
\value{
A \link{Schema}
}
\description{
read a Schema from a stream
}
