% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatch.R
\name{read_stream}
\alias{read_stream}
\title{Reads a stream of arrow::RecordBatch from a raw() vector}
\usage{
read_stream(stream)
}
\arguments{
\item{stream}{a \code{raw()} containing a stream of record batches.}
}
\description{
Reads a stream of arrow::RecordBatch from a raw() vector
}
