% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\name{read_table}
\alias{read_table}
\title{Read an arrow::Table from a stream}
\usage{
read_table(stream)
}
\arguments{
\item{stream}{stream. Either a stream created by \code{\link[=file_open]{file_open()}} or \code{\link[=mmap_open]{mmap_open()}} or a file path.}
}
\description{
Read an arrow::Table from a stream
}
