% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadableFile}
\alias{ReadableFile}
\title{open a \link[=arrow__io__ReadableFile]{arrow::io::ReadableFile}}
\usage{
ReadableFile(path)
}
\arguments{
\item{path}{file path}
}
\value{
a \link[=arrow__io__ReadableFile]{arrow::io::ReadableFile}
}
\description{
open a \link[=arrow__io__ReadableFile]{arrow::io::ReadableFile}
}
