% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\name{record_batch_file_reader}
\alias{record_batch_file_reader}
\title{Create an \code{arrow::ipc::RecordBatchFileReader} from a file}
\usage{
record_batch_file_reader(file)
}
\arguments{
\item{file}{The file to read from}
}
\description{
Create an \code{arrow::ipc::RecordBatchFileReader} from a file
}
