% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{record_batch_file_writer}
\alias{record_batch_file_writer}
\title{Create a record batch file writer from a stream}
\usage{
record_batch_file_writer(stream, schema)
}
\arguments{
\item{stream}{a stream}

\item{schema}{the schema of the batches}
}
\value{
an \code{arrow::ipc::RecordBatchWriter} object
}
\description{
Create a record batch file writer from a stream
}
