% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\name{record_batch_stream_reader}
\alias{record_batch_stream_reader}
\title{Create a \code{arrow::ipc::RecordBatchStreamReader} from an input stream}
\usage{
record_batch_stream_reader(stream)
}
\arguments{
\item{stream}{input stream}
}
\description{
Create a \code{arrow::ipc::RecordBatchStreamReader} from an input stream
}
