% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{record_batch_stream_writer}
\alias{record_batch_stream_writer}
\title{Create a record batch writer that writes to the given stream}
\usage{
record_batch_stream_writer(stream, schema)
}
\arguments{
\item{stream}{an stream}

\item{schema}{a Schema}
}
\description{
Create a record batch writer that writes to the given stream
}
