% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-batch-reader.R
\docType{class}
\name{RecordBatchFileReader}
\alias{RecordBatchFileReader}
\title{class arrow::RecordBatchFileReader}
\description{
class arrow::RecordBatchFileReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
