% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-batch-reader.R
\docType{class}
\name{RecordBatchReader}
\alias{RecordBatchReader}
\title{class arrow::RecordBatchReader}
\description{
class arrow::RecordBatchReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
