% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-batch-reader.R
\docType{class}
\name{RecordBatchStreamReader}
\alias{RecordBatchStreamReader}
\title{class arrow::RecordBatchStreamReader}
\description{
class arrow::RecordBatchStreamReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
