% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\docType{class}
\name{RecordBatchWriter}
\alias{RecordBatchWriter}
\title{class arrow::RecordBatchWriter}
\description{
class arrow::RecordBatchWriter
}
\section{Methods}{

\itemize{
\item \code{$write_batch(batch)}: Write record batch to stream
\item \code{$write_table(table)}: write Table to stream
\item \code{$close()}: close stream
}
}

\section{Derived classes}{

\itemize{
\item \link[=RecordBatchStreamWriter]{arrow::RecordBatchStreamWriter} implements the streaming binary format
\item \link[=RecordBatchFileWriter]{arrow::RecordBatchFileWriter} implements the binary file format
}
}

\keyword{datasets}
