% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recycle_scalars}
\alias{recycle_scalars}
\title{Recycle scalar values in a list of arrays}
\usage{
recycle_scalars(arrays)
}
\arguments{
\item{arrays}{List of arrays}
}
\value{
List of arrays with any vector/Scalar/Array/ChunkedArray values of length 1 recycled
}
\description{
Recycle scalar values in a list of arrays
}
\keyword{internal}
