% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{repeat_value_as_array}
\alias{repeat_value_as_array}
\title{Take an object of length 1 and repeat it.}
\usage{
repeat_value_as_array(object, n)
}
\arguments{
\item{object}{Object of length 1 to be repeated - vector, \code{Scalar}, \code{Array}, or \code{ChunkedArray}}

\item{n}{Number of repetitions}
}
\value{
\code{Array} of length \code{n}
}
\description{
Take an object of length 1 and repeat it.
}
\keyword{internal}
