% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate-round.R
\name{round_temporal}
\alias{round_temporal}
\title{Round a temporal object}
\usage{
round_temporal(x, unit = "second")
}
\arguments{
\item{x}{Scalar or Array to be rounded}

\item{unit}{string specifying the calendar unit (default = "second")}
}
\value{
Scalar or Array containing rounded times
}
\description{
Round a temporal object
}
\examples{
timey_wimeys <- Array$create(Sys.time() + rnorm(10, sd = 360))
round_temporal(timey_wimeys)
round_temporal(timey_wimeys, unit = "minute")
}
