% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{scalar}
\alias{scalar}
\title{Create an Arrow Scalar}
\usage{
scalar(x, type = NULL)
}
\description{
Create an Arrow Scalar
}
\examples{
scalar(pi)
scalar(404)
# If you pass a vector into scalar(), you get a list containing your items
scalar(c(1, 2, 3))

scalar(9) == scalar(10)

}
