% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_cpu_count}
\alias{set_cpu_count}
\title{Set number of threads used by the global CPU thread pool in libarrow}
\usage{
set_cpu_count(num_threads)
}
\arguments{
\item{num_threads}{integer: New number of threads for thread pool}
}
\description{
Set number of threads used by the global CPU thread pool in libarrow
}
