% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetCpuThreadPoolCapacity}
\alias{SetCpuThreadPoolCapacity}
\title{Set the capacity of the global thread pool}
\usage{
SetCpuThreadPoolCapacity(threads)
}
\arguments{
\item{threads}{the number of worker threads int the thread pool to which
Arrow dispatches various CPU-bound tasks.

The current number is returned by \code{\link[=GetCpuThreadPoolCapacity]{GetCpuThreadPoolCapacity()}}}
}
\description{
Set the capacity of the global thread pool
}
