% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{split_pattern}
\alias{split_pattern}
\title{String splitting}
\usage{
split_pattern(x, pattern, max_splits = -1L, reverse = FALSE)
}
\arguments{
\item{x}{\code{Scalar}, \code{Array} or \code{ChunkedArray}}

\item{pattern}{The exact substring to look for inside input values.}

\item{max_splits}{Maximum number of splits. Default -1 means no limit.}

\item{reverse}{When true, the splitting is done from the end of the string.}
}
\value{
An object (same type as \code{x}) containing lists of strings
}
\description{
Split a string into a list of strings
}
