% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R, R/RecordBatchWriter.R
\name{stream}
\alias{stream}
\alias{stream.arrow::RecordBatch}
\alias{stream.arrow::Table}
\title{Stream an arrow object}
\usage{
stream(x, stream, ...)

\method{"stream"}{"arrow::RecordBatch"}(x, stream, ...)

\method{"stream"}{"arrow::Table"}(x, stream, ...)
}
\arguments{
\item{x}{an arrow object}

\item{stream}{an output stream}

\item{...}{Additional parameters}
}
\description{
Stream an arrow object
}
