% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{stream_RecordBatch}
\alias{stream_RecordBatch}
\title{Stream a record batch}
\usage{
stream_RecordBatch(x, stream, ...)
}
\arguments{
\item{x}{an arrow object}

\item{stream}{an output stream}

\item{...}{Additional parameters}
}
\description{
Stream a record batch
}
