% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{stream_Table}
\alias{stream_Table}
\title{stream an arrow::Table}
\usage{
stream_Table(x, stream, ...)
}
\arguments{
\item{x}{an arrow object}

\item{stream}{an output stream}

\item{...}{Additional parameters}
}
\description{
stream an arrow::Table
}
