% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{table_from_batches}
\alias{table_from_batches}
\title{Create a arrow::Table from record batches}
\usage{
table_from_batches(..., schema = NULL)
}
\arguments{
\item{...}{variable number of arrow::RecordBatch. This supports tidy dots splicing}

\item{schema}{either NULL or a arrow::Schema}
}
\value{
a arrow::Table
}
\description{
Create a arrow::Table from record batches
}
