% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-arrow.R
\name{table_from_dots}
\alias{table_from_dots}
\title{Create an arrow::Table from diverse inputs}
\usage{
table_from_dots(..., schema = NULL)
}
\arguments{
\item{...}{arrays, chunked arrays, or R vectors that should define the
columns of the Arrow Table; alternatively, if record batches are given,
they will be stacked.}

\item{schema}{a schema. The default (\code{NULL}) infers the schema from the \code{...}}
}
\value{
An \code{arrow::Table}
}
\description{
Unlike \code{\link[=to_arrow]{to_arrow()}}, this function splices together inputs to form a Table.
When providing columns, they can be a mix of Arrow arrays and R vectors.
}
\examples{
\donttest{
tab1 <- table_from_dots(a = 1:10, b = letters[1:10])
tab1
as.data.frame(tab1)
}
}
