% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{table_writer}
\alias{table_writer}
\title{Create TableWriter that writes into a stream}
\usage{
table_writer(stream)
}
\arguments{
\item{stream}{an OutputStream}
}
\description{
Create TableWriter that writes into a stream
}
