% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threadpool}
\alias{threadpool}
\alias{GetCpuThreadPoolCapacity}
\alias{SetCpuThreadPoolCapacity}
\title{View and manage the capacity of the global thread pool}
\usage{
GetCpuThreadPoolCapacity()

SetCpuThreadPoolCapacity(threads)
}
\arguments{
\item{threads}{the number of worker threads in the thread pool to which
Arrow dispatches various CPU-bound tasks.}
}
\value{
\code{GetCpuThreadPoolCapacity()} returns the number of worker threads.
\code{SetCpuThreadPoolCapacity()} returns nothing.
}
\description{
\code{GetCpuThreadPoolCapacity()} returns the number of worker threads in the
thread pool to which
Arrow dispatches various CPU-bound tasks. This is an ideal number,
not necessarily the exact number of threads at a given point in time.
You can change this number using \code{SetCpuThreadPoolCapacity()}.
}
