% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{to_duckdb}
\alias{to_duckdb}
\alias{to_duckdb.Dataset}
\alias{to_duckdb.Table}
\alias{to_duckdb.arrow_dplyr_query}
\title{Create a (virtual) DuckDB table from an Arrow object}
\usage{
to_duckdb(src, con, table_name, auto_disconnect)

\method{to_duckdb}{Dataset}(
  src,
  con = arrow_duck_connection(),
  table_name = unique_arrow_tablename(),
  auto_disconnect = TRUE
)

\method{to_duckdb}{Table}(
  src,
  con = arrow_duck_connection(),
  table_name = unique_arrow_tablename(),
  auto_disconnect = TRUE
)

\method{to_duckdb}{arrow_dplyr_query}(
  src,
  con = arrow_duck_connection(),
  table_name = unique_arrow_tablename(),
  auto_disconnect = TRUE
)
}
\arguments{
\item{src}{the Arrow object (e.g. Dataset, Table) to use for the DuckDB table}

\item{con}{a DuckDB connection to use (default will create one and store it
in \code{options("arrow_duck_con")})}

\item{table_name}{a name to use in DuckDB for this object. The default is a
unique string \code{"arrow_"} followed by numbers.}

\item{auto_disconnect}{should the table be automatically cleaned up when the
resulting object is removed (and garbage collected)? Default: \code{TRUE}}
}
\value{
A tbl of the new table in DuckDB
}
\description{
This will do the necessary configuration to create a (virtual) table in DuckDB
that is backed by the Arrow object given. No data is copied or modified until
\code{collect()} or \code{compute()} are called or a query is run against the table.
}
\details{
The result is a dbplyr-compatible object that can be used in d(b)plyr pipelines.
}
\keyword{internal}
