% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpch.R
\name{tpch_dbgen}
\alias{tpch_dbgen}
\title{Generate a RecordBatchReader with TPC-H data in it}
\usage{
tpch_dbgen(table = tpch_tables, scale_factor)
}
\arguments{
\item{table}{the table to generate}

\item{scale_factor}{the scale factor to generate}
}
\value{
a RecordBatchReader that will contain the generated data
}
\description{
Generate a RecordBatchReader with TPC-H data in it
}
\keyword{internal}
