% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpch.R
\name{tpch_dbgen_write}
\alias{tpch_dbgen_write}
\title{Generate TPC-H data as datasets}
\usage{
tpch_dbgen_write(scale_factor, path, ...)
}
\arguments{
\item{scale_factor}{numeric, scale factor of data to generate}

\item{...}{}
}
\value{
the path to the directory containing the datasets
}
\description{
This will generate all of the TPC-H data as a set of datasets for a given
\code{scale_factor}
}
\keyword{internal}
