% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{unify_schemas}
\alias{unify_schemas}
\title{Combine and harmonize schemas}
\usage{
unify_schemas(..., schemas = list(...))
}
\arguments{
\item{...}{\link{Schema}s to unify}

\item{schemas}{Alternatively, a list of schemas}
}
\value{
A \code{Schema} with the union of fields contained in the inputs
}
\description{
Combine and harmonize schemas
}
\examples{
\dontrun{
a <- schema(b = double(), c = bool())
z <- schema(b = double(), k = utf8())
unify_schemas(a, z)
}
}
