% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{use_run_with_captured_r}
\alias{use_run_with_captured_r}
\alias{set_use_run_with_captured_r}
\title{Manage experimental R code evaluation from within Arrow}
\usage{
use_run_with_captured_r()

set_use_run_with_captured_r(run_with_captured_r = TRUE)
}
\arguments{
\item{run_with_captured_r}{Use TRUE to enable R code evaluation from within
arrow}
}
\description{
Manage experimental R code evaluation from within Arrow
}
