% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{value_counts}
\alias{value_counts}
\title{\code{table} for Arrow objects}
\usage{
value_counts(x)
}
\arguments{
\item{x}{\code{Array} or \code{ChunkedArray}}
}
\value{
A \code{StructArray} containing "values" (same type as \code{x}) and "counts"
\code{Int64}.
}
\description{
This function tabulates the values in the array and returns a table of counts.
}
\examples{
cyl_vals <- Array$create(mtcars$cyl)
counts <- value_counts(cyl_vals)
}
