% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_arrow.arrow::Table}
\alias{write_arrow.arrow::Table}
\title{#' @export
\code{write_arrow.arrow::RecordBatch} <- function(x, stream, ...) {
write_record_batch(x, stream, ...)
}}
\usage{
\method{write_arrow}{arrow::Table}(x, stream, ...)
}
\description{
#' @export
\code{write_arrow.arrow::RecordBatch} <- function(x, stream, ...) {
write_record_batch(x, stream, ...)
}
}
