% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-write.R
\name{write_csv_dataset}
\alias{write_csv_dataset}
\title{Write a dataset into partitioned csv files.}
\usage{
write_csv_dataset(
  dataset,
  path,
  partitioning = dplyr::group_vars(dataset),
  basename_template = "part-{i}.csv",
  hive_style = TRUE,
  existing_data_behavior = c("overwrite", "error", "delete_matching"),
  max_partitions = 1024L,
  max_open_files = 900L,
  max_rows_per_file = 0L,
  min_rows_per_group = 0L,
  max_rows_per_group = bitwShiftL(1, 20),
  ...
)
}
\description{
A wrapper around \link{write_dataset} to allow for easy switching between functions for writing datasets.
}
\seealso{
\code{\link[=write_dataset]{write_dataset()}}
}
