% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{write_feather_RecordBatch}
\alias{write_feather_RecordBatch}
\title{Write a record batch in the feather format}
\usage{
write_feather_RecordBatch(data, stream)
}
\arguments{
\item{data}{\code{data.frame} or \code{arrow::RecordBatch}}

\item{stream}{A file path or an \code{arrow::io::OutputStream}}
}
\description{
Write a record batch in the feather format
}
\keyword{internal}
