% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/ipc_stream.R
\name{write_arrow}
\alias{write_arrow}
\alias{write_ipc_stream}
\title{Write Arrow IPC stream format}
\usage{
write_arrow(x, sink, ...)

write_ipc_stream(x, sink, ...)
}
\arguments{
\item{x}{\code{data.frame}, \link{RecordBatch}, or \link{Table}}

\item{sink}{A string file path, URI, or \link{OutputStream}, or path in a file
system (\code{SubTreeFileSystem})}

\item{...}{extra parameters passed to \code{write_feather()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Apache Arrow defines two formats for \href{https://arrow.apache.org/docs/format/Columnar.html#serialization-and-interprocess-communication-ipc}{serializing data for interprocess communication (IPC)}:
a "stream" format and a "file" format, known as Feather. \code{write_ipc_stream()}
and \code{\link[=write_feather]{write_feather()}} write those formats, respectively.
}
\details{
\code{write_arrow()}, a wrapper around \code{write_ipc_stream()} and \code{write_feather()}
with some nonstandard behavior, is deprecated. You should explicitly choose
the function that will write the desired IPC format (stream or file) since
either can be written to a file or \code{OutputStream}.
}
\examples{
\dontshow{if (arrow_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tf <- tempfile()
on.exit(unlink(tf))
write_ipc_stream(mtcars, tf)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=write_feather]{write_feather()}} for writing IPC files. \code{\link[=write_to_raw]{write_to_raw()}} to
serialize data to a buffer.
\link{RecordBatchWriter} for a lower-level interface.
}
