% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{write_record_batch}
\alias{write_record_batch}
\title{write a record batch}
\usage{
write_record_batch(x, stream, ...)
}
\arguments{
\item{x}{a \code{arrow::RecordBatch}}

\item{stream}{where to stream the record batch}

\item{...}{extra parameters}
}
\description{
write a record batch
}
