% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{write_table}
\alias{write_table}
\title{write an arrow::Table}
\usage{
write_table(x, stream, ...)
}
\arguments{
\item{x}{an \code{arrow::Table}}

\item{stream}{where to stream the record batch}

\item{...}{extra parameters}
}
\description{
write an arrow::Table
}
