% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-funcs-augmented.R
\name{add_filename}
\alias{add_filename}
\title{Add the data filename as a column}
\usage{
add_filename()
}
\value{
A \code{FieldRef} \code{\link{Expression}} that refers to the filename
augmented column.
}
\description{
This function only exists inside \code{arrow} \code{dplyr} queries, and it only is
valid when querying on a \code{FileSystemDataset}, such as one created by
\code{\link[=open_dataset]{open_dataset()}}. Use it inside \code{mutate()} to add a column holding the path
of the file each row was read from.
}
\details{
To use filenames generated by this function in subsequent pipeline steps, you
must either call \code{\link[dplyr:compute]{compute()}} or
\code{\link[dplyr:collect]{collect()}} first. It must also be added before
any aggregation or join. See Examples.
}
\examples{
\dontrun{
open_dataset("nyc-taxi") |>
  mutate(file = add_filename()) |>
  collect()

# To refer to the new column in a later step of the same query (for
# example, to extract partition values from the path), call compute() or
# collect() first
open_dataset("nyc-taxi") |>
  mutate(file = add_filename()) |>
  compute() |>
  mutate(filename_length = nchar(file))
}

}
\seealso{
\code{\link[=open_dataset]{open_dataset()}}
}
\keyword{internal}
