% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-funcs-augmented.R
\name{add_filename}
\alias{add_filename}
\title{Add the data filename as a column}
\usage{
add_filename()
}
\value{
A \code{FieldRef} \code{\link{Expression}} that refers to the filename
augmented column.
}
\description{
This function only exists inside \code{arrow} \code{dplyr} queries, and it only is
valid when querying on a \code{FileSystemDataset}, such as one created by
\code{\link[=open_dataset]{open_dataset()}}. Use it inside \code{mutate()} to add a column holding the path
of the file each row was read from.
}
\details{
The filename column can be used in later \code{mutate()}, \code{select()}, \code{arrange()}
and \code{group_by()} steps of the same query, but it cannot be used in a
\code{filter()} until you have called \code{\link[dplyr:compute]{compute()}} or
\code{\link[dplyr:collect]{collect()}}. It must also be added before any
aggregation or join. See Examples.
}
\examples{
\dontrun{
open_dataset("nyc-taxi") |>
  mutate(file = add_filename()) |>
  collect()

# The new column can be used in a later mutate() in the same query
open_dataset("nyc-taxi") |>
  mutate(file = add_filename()) |>
  mutate(filename_length = nchar(file)) |>
  collect()

# To filter() on the new column, call compute() or collect() first
open_dataset("nyc-taxi") |>
  mutate(file = add_filename()) |>
  compute() |>
  filter(endsWith(file, "part-0.parquet")) |>
  collect()
}

}
\seealso{
\code{\link[=open_dataset]{open_dataset()}}
}
\keyword{internal}
