% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_filename}
\alias{add_filename}
\title{Add filename as dataset column}
\usage{
add_filename()
}
\description{
Add filename as dataset column
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(dplyr)
tf <- tempfile()
dir.create(tf)
write_dataset(mtcars, tf, partitioning = "cyl")

open_dataset(tf) \%>\%
 mutate(filename = add_filename()) \%>\%
 collect()
\dontshow{\}) # examplesIf}
}
