% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-funcs-augmented.R
\name{add_filename}
\alias{add_filename}
\title{Add the data filename as a column}
\usage{
add_filename()
}
\value{
A \code{FieldRef} \code{Expression} that refers to the filename augmented
column.
}
\description{
This function only exists inside \code{arrow} \code{dplyr} queries, and it only is
valid when quering on a \code{FileSystemDataset}.
}
\examples{
\dontrun{
open_dataset("nyc-taxi") \%>\%
  mutate(file = add_filename())
}
}
\keyword{internal}
