% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{array}
\alias{array}
\title{create an \link[=arrow__Array]{arrow::Array} from an R vector}
\usage{
array(..., type)
}
\arguments{
\item{\dots}{Vectors to coerce}

\item{type}{currently ignored}
}
\description{
create an \link[=arrow__Array]{arrow::Array} from an R vector
}
