% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{array}
\alias{array}
\title{create an \link[=arrow__Array]{arrow::Array} from an R vector}
\usage{
array(x, type = NULL)
}
\arguments{
\item{x}{R object}

\item{type}{Explicit \link[=arrow__DataType]{type}, or NULL (the default) to infer from the data}
}
\description{
create an \link[=arrow__Array]{arrow::Array} from an R vector
}
