% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-package.R
\docType{package}
\name{arrow-package}
\alias{arrow}
\alias{arrow-package}
\title{arrow: Integration to 'Apache' 'Arrow'}
\description{
'Apache' 'Arrow' <https://arrow.apache.org/> is a cross-language
    development platform for in-memory data. It specifies a standardized
    language-independent columnar memory format for flat and hierarchical data,
    organized for efficient analytic operations on modern hardware. This
    package provides an interface to the Arrow C++ library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arrow.apache.org/docs/r/}
  \item \url{https://github.com/apache/arrow/}
  \item Report bugs at \url{https://issues.apache.org/jira/projects/ARROW/issues}
}

}
\author{
\strong{Maintainer}: Romain François \email{romain@rstudio.com}

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroen@berkeley.edu}
  \item Apache Arrow \email{dev@arrow.apache.org} [copyright holder]
}

Other contributors:
\itemize{
  \item Javier Luraschi \email{javier@rstudio.com} [contributor]
  \item Jeffrey Wong \email{jeffreyw@netflix.com} [contributor]
}

}
\keyword{internal}
