% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__BufferReader}
\alias{arrow__io__BufferReader}
\alias{arrow::io::BufferReader}
\title{class arrow::io::BufferReader}
\description{
class arrow::io::BufferReader
}
\section{Methods}{


TODO
}

\keyword{datasets}
