% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__FileOutputStream}
\alias{arrow__io__FileOutputStream}
\alias{arrow::io::FileOutputStream}
\title{class arrow::io::FileOutputStream}
\description{
class arrow::io::FileOutputStream
}
\section{Methods}{


TODO
}

\keyword{datasets}
