% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__FixedSizeBufferWriter}
\alias{arrow__io__FixedSizeBufferWriter}
\alias{arrow::io::FixedSizeBufferWriter}
\title{class arrow::io::FixedSizeBufferWriter}
\description{
class arrow::io::FixedSizeBufferWriter
}
\section{Methods}{


TODO
}

\keyword{datasets}
