% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__InputStream}
\alias{arrow__io__InputStream}
\alias{arrow::io::InputStream}
\title{class arrow::io::InputStream}
\description{
class arrow::io::InputStream
}
\section{Methods}{


TODO
}

\keyword{datasets}
