% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__MemoryMappedFile}
\alias{arrow__io__MemoryMappedFile}
\alias{arrow::io::MemoryMappedFile}
\title{class arrow::io::MemoryMappedFile}
\description{
class arrow::io::MemoryMappedFile
}
\section{Methods}{


TODO
}

\seealso{
\code{\link[=mmap_open]{mmap_open()}}, \code{\link[=mmap_create]{mmap_create()}}
}
\keyword{datasets}
