% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__MockOutputStream}
\alias{arrow__io__MockOutputStream}
\alias{arrow::io::MockOutputStream}
\title{class arrow::io::MockOutputStream}
\description{
class arrow::io::MockOutputStream
}
\section{Methods}{


TODO
}

\keyword{datasets}
