% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__Readable}
\alias{arrow__io__Readable}
\alias{arrow::io::Readable}
\title{class arrow::io::Readable}
\description{
class arrow::io::Readable
}
\section{Methods}{


TODO
}

\keyword{datasets}
