% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{arrow__io__ReadableFile}
\alias{arrow__io__ReadableFile}
\alias{arrow::io::ReadableFile}
\title{class arrow::io::ReadableFile}
\description{
class arrow::io::ReadableFile
}
\section{Methods}{


TODO
}

\keyword{datasets}
