% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arrow_array}
\alias{arrow_array}
\title{Create an Arrow Array}
\usage{
arrow_array(x, type = NULL)
}
\arguments{
\item{x}{An R object representable as an Arrow array, e.g. a vector, list, or \code{data.frame}.}

\item{type}{An optional \link[=data-type]{data type} for \code{x}. If omitted, the type will be inferred from the data.}
}
\description{
Create an Arrow Array
}
\examples{
my_array <- arrow_array(1:10)

# Compare 2 arrays
na_array <- arrow_array(c(1:5, NA))
na_array2 <- na_array
na_array2 == na_array # element-wise comparison
}
