% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-package.R
\name{arrow_available}
\alias{arrow_available}
\alias{arrow_with_dataset}
\alias{arrow_with_engine}
\alias{arrow_with_parquet}
\alias{arrow_with_s3}
\alias{arrow_with_json}
\title{Is the C++ Arrow library available?}
\usage{
arrow_available()

arrow_with_dataset()

arrow_with_engine()

arrow_with_parquet()

arrow_with_s3()

arrow_with_json()
}
\value{
\code{TRUE} or \code{FALSE} depending on whether the package was installed
with:
\itemize{
\item The Arrow C++ library (check with \code{arrow_available()})
\item Arrow Dataset support enabled (check with \code{arrow_with_dataset()})
\item Parquet support enabled (check with \code{arrow_with_parquet()})
\item JSON support enabled (check with \code{arrow_with_json()})
\item Amazon S3 support enabled (check with \code{arrow_with_s3()})
}
}
\description{
You won't generally need to call these function, but they're made available
for diagnostic purposes.
}
\examples{
arrow_available()
arrow_with_dataset()
arrow_with_parquet()
arrow_with_json()
arrow_with_s3()
}
\seealso{
If any of these are \code{FALSE}, see
\code{vignette("install", package = "arrow")} for guidance on reinstalling the
package.
}
