% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-info.R
\name{arrow_info}
\alias{arrow_info}
\alias{arrow_available}
\alias{arrow_with_dataset}
\alias{arrow_with_substrait}
\alias{arrow_with_parquet}
\alias{arrow_with_s3}
\alias{arrow_with_gcs}
\alias{arrow_with_json}
\title{Report information on the package's capabilities}
\usage{
arrow_info()

arrow_available()

arrow_with_dataset()

arrow_with_substrait()

arrow_with_parquet()

arrow_with_s3()

arrow_with_gcs()

arrow_with_json()
}
\value{
\code{arrow_info()} returns a list including version information, boolean
"capabilities", and statistics from Arrow's memory allocator, and also
Arrow's run-time information. The \verb{_available()} functions return a logical
value whether or not the C++ library was built with support for them.
}
\description{
This function summarizes a number of build-time configurations and run-time
settings for the Arrow package. It may be useful for diagnostics.
}
\seealso{
If any capabilities are \code{FALSE}, see the
\href{https://arrow.apache.org/docs/r/articles/install.html}{install guide}
for guidance on reinstalling the package.
}
