% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-batch-reader.R
\name{as_record_batch_reader}
\alias{as_record_batch_reader}
\alias{as_record_batch_reader.RecordBatchReader}
\alias{as_record_batch_reader.Table}
\alias{as_record_batch_reader.RecordBatch}
\alias{as_record_batch_reader.data.frame}
\alias{as_record_batch_reader.Dataset}
\alias{as_record_batch_reader.function}
\alias{as_record_batch_reader.arrow_dplyr_query}
\alias{as_record_batch_reader.Scanner}
\title{Convert an object to an Arrow RecordBatchReader}
\usage{
as_record_batch_reader(x, ...)

\method{as_record_batch_reader}{RecordBatchReader}(x, ...)

\method{as_record_batch_reader}{Table}(x, ...)

\method{as_record_batch_reader}{RecordBatch}(x, ...)

\method{as_record_batch_reader}{data.frame}(x, ...)

\method{as_record_batch_reader}{Dataset}(x, ...)

\method{as_record_batch_reader}{`function`}(x, ..., schema)

\method{as_record_batch_reader}{arrow_dplyr_query}(x, ...)

\method{as_record_batch_reader}{Scanner}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \link{RecordBatchReader}}

\item{...}{Passed to S3 methods}

\item{schema}{The \code{\link[=schema]{schema()}} that must match the schema returned by each
call to \code{x} when \code{x} is a function.}
}
\value{
A \link{RecordBatchReader}
}
\description{
Convert an object to an Arrow RecordBatchReader
}
\examples{
\dontshow{if (arrow_with_dataset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
reader <- as_record_batch_reader(data.frame(col1 = 1, col2 = "two"))
reader$read_next_batch()
\dontshow{\}) # examplesIf}
}
