% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\title{Create a buffer from an R object}
\usage{
buffer(x)
}
\arguments{
\item{x}{R object}
}
\value{
an instance of \code{arrow::Buffer} that borrows memory from \code{x}
}
\description{
Create a buffer from an R object
}
