% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\title{Create a \link[=arrow__Buffer]{arrow::Buffer} from an R object}
\usage{
buffer(x)
}
\arguments{
\item{x}{R object. Only raw, numeric and integer vectors are currently supported}
}
\value{
an instance of \link[=arrow__Buffer]{arrow::Buffer} that borrows memory from \code{x}
}
\description{
Create a \link[=arrow__Buffer]{arrow::Buffer} from an R object
}
