% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{BufferOutputStream}
\alias{BufferOutputStream}
\title{Open a \link[=arrow__io__BufferOutputStream]{arrow::io::BufferOutputStream}}
\usage{
BufferOutputStream(initial_capacity = 0L)
}
\arguments{
\item{initial_capacity}{initial capacity}
}
\value{
a \link[=arrow__io__BufferOutputStream]{arrow::io::BufferOutputStream}
}
\description{
Open a \link[=arrow__io__BufferOutputStream]{arrow::io::BufferOutputStream}
}
