% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{cast_options}
\alias{cast_options}
\title{Cast options}
\usage{
cast_options(safe = TRUE, allow_int_overflow = !safe,
  allow_time_truncate = !safe, allow_float_truncate = !safe)
}
\arguments{
\item{safe}{enforce safe conversion}

\item{allow_int_overflow}{allow int conversion, \code{!safe} by default}

\item{allow_time_truncate}{allow time truncate, \code{!safe} by default}

\item{allow_float_truncate}{allow float truncate, \code{!safe} by default}
}
\description{
Cast options
}
