% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkedArray.R
\name{chunked_array}
\alias{chunked_array}
\title{create an arrow::Array from an R vector}
\usage{
chunked_array(...)
}
\arguments{
\item{\dots}{Vectors to coerce}
}
\description{
create an arrow::Array from an R vector
}
