% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkedArray.R
\name{chunked_array}
\alias{chunked_array}
\title{create an \link[=arrow__ChunkedArray]{arrow::ChunkedArray} from various R vectors}
\usage{
chunked_array(..., type)
}
\arguments{
\item{\dots}{Vectors to coerce}

\item{type}{currently ignored}
}
\description{
create an \link[=arrow__ChunkedArray]{arrow::ChunkedArray} from various R vectors
}
