% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\docType{class}
\name{Codec}
\alias{Codec}
\title{Compression Codec class}
\description{
Codecs allow you to create \link[=compression]{compressed input and output streams}.
}
\section{Factory}{

The \code{Codec$create()} factory method takes the following arguments:
\itemize{
\item \code{type}: string name of the compression method. Possible values are
"uncompressed", "snappy", "gzip", "brotli", "zstd", "lz4", "lzo", or
"bz2". \code{type} may be upper- or lower-cased. Not all methods may be
available; support depends on build-time flags for the C++ library.
See \code{\link[=codec_is_available]{codec_is_available()}}. Most builds support at least "snappy" and
"gzip". All support "uncompressed".
\item \code{compression_level}: compression level, the default value (\code{NA}) uses the
default compression level for the selected compression \code{type}.
}
}

