% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\docType{class}
\name{compression}
\alias{compression}
\alias{CompressedOutputStream}
\alias{CompressedInputStream}
\title{Compressed stream classes}
\description{
\code{CompressedInputStream} and \code{CompressedOutputStream}
allow you to apply a \code{\link[=compression_codec]{compression_codec()}} to an
input or output stream.
}
\section{Factory}{


The \code{CompressedInputStream$create()} and \code{CompressedOutputStream$create()}
factory methods instantiate the object and take the following arguments:
\itemize{
\item \code{stream} An \link{InputStream} or \link{OutputStream}, respectively
\item \code{codec} A \code{Codec}
}
}

\section{Methods}{


Methods are inherited from \link{InputStream} and \link{OutputStream}, respectively
}

\keyword{datasets}
