% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cpu_count}
\alias{cpu_count}
\alias{set_cpu_count}
\title{Manage the global CPU thread pool in libarrow}
\usage{
cpu_count()

set_cpu_count(num_threads)
}
\arguments{
\item{num_threads}{integer: New number of threads for thread pool}
}
\description{
Manage the global CPU thread pool in libarrow
}
