% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{csv_convert_options}
\alias{csv_convert_options}
\title{Conversion Options for the csv reader}
\usage{
csv_convert_options(check_utf8 = TRUE)
}
\arguments{
\item{check_utf8}{Whether to check UTF8 validity of string columns}
}
\description{
Conversion Options for the csv reader
}
