% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{csv_read_options}
\alias{csv_read_options}
\title{read options for the csv reader}
\usage{
csv_read_options(block_size = 1048576L)
}
\arguments{
\item{block_size}{Block size we request from the IO layer; also determines the size of chunks when use_threads is \code{TRUE}}
}
\description{
read options for the csv reader
}
