% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R, R/json.R
\name{csv_read_options}
\alias{csv_read_options}
\alias{json_read_options}
\title{Read options for the Arrow file readers}
\usage{
csv_read_options(use_threads = option_use_threads(),
  block_size = 1048576L, skip_rows = 0L, column_names = character(0),
  autogenerate_column_names = FALSE)

json_read_options(use_threads = TRUE, block_size = 1048576L)
}
\arguments{
\item{use_threads}{Whether to use the global CPU thread pool}

\item{block_size}{Block size we request from the IO layer; also determines
the size of chunks when use_threads is \code{TRUE}. NB: if \code{FALSE}, JSON input
must end with an empty line.}

\item{skip_rows}{Number of lines to skip before reading data.}

\item{column_names}{Character vector to supply column names. If length-0
(the default), the first non-skipped row will be parsed to generate column
names, unless \code{autogenerate_column_names} is \code{TRUE}.}

\item{autogenerate_column_names}{Logical: generate column names instead of
using the first non-skipped row (the default)? If \code{TRUE}, column names will
be "f0", "f1", ..., "fN".}
}
\description{
Read options for the Arrow file readers
}
