% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{csv_table_reader}
\alias{csv_table_reader}
\title{CSV table reader}
\usage{
csv_table_reader(file, read_options = csv_read_options(),
  parse_options = csv_parse_options(),
  convert_options = csv_convert_options(), ...)
}
\arguments{
\item{file}{file}

\item{read_options, }{see \code{\link[=csv_read_options]{csv_read_options()}}}

\item{parse_options, }{see \code{\link[=csv_parse_options]{csv_parse_options()}}}

\item{convert_options, }{see \code{\link[=csv_convert_options]{csv_convert_options()}}}

\item{...}{additional parameters.}
}
\description{
CSV table reader
}
