% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset}
\alias{Dataset}
\title{Multi-file datasets}
\value{
A \link{ScannerBuilder}
}
\description{
Arrow Datasets allow you to query against data that has been split across
multiple files. This sharding of data may indicate partitioning, which
can accelerate queries that only touch some partitions (files).

Start a new scan of the data

Return the Dataset's \code{Schema}
}
\section{Factory}{

The \code{Dataset$create()} factory method instantiates a \code{Dataset} and
takes the following arguments:
\itemize{
\item \code{sources}: a list of \link{DataSource} objects
\item \code{schema}: a \link{Schema}
}
}

\section{Methods}{

\itemize{
\item \verb{$NewScan()}: Returns a \link{ScannerBuilder} for building a query
\item \verb{$schema}: Active binding, returns the \link{Schema} of the Dataset
}
}

\seealso{
\code{\link[=open_dataset]{open_dataset()}} for a simple way to create a Dataset that has a
single \code{DataSource}.
}
