% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{DataSource}
\alias{DataSource}
\alias{DataSourceDiscovery}
\alias{FileSystemDataSourceDiscovery}
\title{Data sources for a Dataset}
\description{
A \link{Dataset} can have one or more \code{DataSource}s. A \code{DataSource} contains one
or more \code{DataFragments}, such as files, of a common type and partition
scheme. \code{DataSourceDiscovery} is used to create a \code{DataSource}, inspect the
\link{Schema} of the fragments contained in it, and declare a partition scheme.
\code{FileSystemDataSourceDiscovery} is a subclass of \code{DataSourceDiscovery} for
discovering files in the local file system, the only currently supported
file system.
}
\section{Factory}{

The \code{DataSourceDiscovery$create()} factory method instantiates a
\code{DataSourceDiscovery} and takes the following arguments:
\itemize{
\item \code{path}: A string file path containing data files
\item \code{filesystem}: Currently only "local" is supported
\item \code{format}: Currently only "parquet" is supported
\item \code{allow_non_existent}: logical: is \code{path} allowed to not exist? Default
\code{FALSE}. See \link{Selector}.
\item \code{recursive}: logical: should files be discovered in subdirectories of
\item \code{path}? Default \code{TRUE}.
\item \code{...} Additional arguments passed to the \link{FileSystem} \verb{$create()} method
}

\code{FileSystemDataSourceDiscovery$create()} is a lower-level factory method and
takes the following arguments:
\itemize{
\item \code{filesystem}: A \link{FileSystem}
\item \code{selector}: A \link{Selector}
\item \code{format}: Currently only "parquet" is supported
}
}

\section{Methods}{

\code{DataSource} has no defined methods. It is just passed to \code{Dataset$create()}.

\code{DataSourceDiscovery} and its subclasses have the following methods:
\itemize{
\item \verb{$Inspect()}: Walks the files in the directory and returns a common \link{Schema}
\item \verb{$SetPartitionScheme(part)}: Takes a \link{PartitionScheme}
\item \verb{$Finish()}: Returns a \code{DataSource}
}
}

\seealso{
\link{Dataset} for what do do with a \code{DataSource}
}
