% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\docType{class}
\name{DataType}
\alias{DataType}
\title{class arrow::DataType}
\description{
class arrow::DataType
}
\section{Methods}{

\itemize{
\item \verb{$ToString()}: String representation of the DataType
\item \verb{$Equals(other)}: Is the DataType equal to \code{other}
\item \verb{$fields()}: The children fields associated with this type
}
}

\seealso{
\code{\link{data-type}}
}
