% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R, R/R6.R, R/List.R, R/Struct.R
\docType{data}
\name{TimeUnit}
\alias{TimeUnit}
\alias{DateUnit}
\alias{Type}
\alias{StatusCode}
\alias{FileMode}
\alias{MessageType}
\alias{CompressionType}
\alias{int8}
\alias{int16}
\alias{int32}
\alias{int64}
\alias{uint8}
\alias{uint16}
\alias{uint32}
\alias{uint64}
\alias{float16}
\alias{float32}
\alias{float64}
\alias{boolean}
\alias{utf8}
\alias{date32}
\alias{date64}
\alias{time32}
\alias{time64}
\alias{null}
\alias{timestamp}
\alias{decimal}
\alias{list_of}
\alias{struct}
\title{Apache Arrow data types}
\format{An object of class \code{arrow::TimeUnit::type} (inherits from \code{arrow-enum}) of length 4.}
\usage{
TimeUnit

DateUnit

Type

StatusCode

FileMode

MessageType

CompressionType

int8()

int16()

int32()

int64()

uint8()

uint16()

uint32()

uint64()

float16()

float32()

float64()

boolean()

utf8()

date32()

date64()

time32(unit)

time64(unit)

null()

timestamp(unit, timezone)

decimal(precision, scale)

list_of(type)

struct(...)
}
\arguments{
\item{unit}{time unit}

\item{timezone}{time zone}

\item{precision}{precision}

\item{scale}{scale}

\item{type}{type}

\item{...}{...}
}
\description{
Apache Arrow data types
}
\keyword{datasets}
