% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_pool.R
\name{default_memory_pool}
\alias{default_memory_pool}
\title{default \link[=arrow__MemoryPool]{arrow::MemoryPool}}
\usage{
default_memory_pool()
}
\value{
the default \link[=arrow__MemoryPool]{arrow::MemoryPool}
}
\description{
default \link[=arrow__MemoryPool]{arrow::MemoryPool}
}
