% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Create a dictionary type}
\usage{
dictionary(index_type, value_type, ordered = FALSE)
}
\arguments{
\item{index_type}{index type, e.g. \code{\link[=int32]{int32()}}}

\item{value_type}{value type, probably \code{\link[=utf8]{utf8()}}}

\item{ordered}{Is this an ordered dictionary ?}
}
\value{
A \link{DictionaryType}
}
\description{
Create a dictionary type
}
\seealso{
\link[=data-type]{Other Arrow data types}
}
