% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{dictionary type factory}
\usage{
dictionary(type, values, ordered = FALSE)
}
\arguments{
\item{type}{indices type, e.g. \code{\link[=int32]{int32()}}}

\item{values}{values array, typically an arrow array of strings}

\item{ordered}{Is this an ordered dictionary}
}
\value{
a \link[=arrow__DictionaryType]{arrow::DictionaryType}
}
\description{
dictionary type factory
}
