% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_optional_dependencies}
\alias{download_optional_dependencies}
\title{Download all optional Arrow dependencies}
\usage{
download_optional_dependencies(deps_dir)
}
\arguments{
\item{deps_dir}{Directory to save files into. Will be created if necessary.}
}
\value{
TRUE/FALSE for whether the downloads were successful

This function is used for setting up an offline build. If it's possible to
download at build time, don't use this function. Instead, let \code{cmake}
download them for you.
These saved files are only used in the build if \code{ARROW_DEPENDENCY_SOURCE}
is \code{BUNDLED} or \code{AUTO}.
https://arrow.apache.org/docs/developers/cpp/building.html#offline-builds

Steps for an offline install with optional dependencies:
\itemize{
\item Install the \code{arrow} package on a computer with internet access
\item Run this function
\item Copy the saved dependency files to a computer without internet access
\item Export the environment variables printed by this function on the computer
without internet access. For example, this function will print
\verb{export ARROW_THRIFT_URL=/path/to/deps_dir/file.tar.gz}
\itemize{
\item These export commands are also saved in \code{DEFINE_ENV_VARS.sh}, in the same
directory
\item You may have to edit the paths if the copied folder is not accessible at
the same location as it was when \code{download_optional_dependencies()} was
run on the internet-connected computer
}
\item Install the \code{arrow} package on the computer without internet access
\item Run \code{\link[=arrow_info]{arrow_info()}} to check installed capabilities
}
}
\description{
Download all optional Arrow dependencies
}
\examples{
\dontrun{
download_optional_dependencies("arrow-thirdparty")
file.exists("arrow-thirdparty/DEFINE_ENV_VARS.sh") # TRUE
list.files("arrow-thirdparty", "thrift-*") # "thrift-0.13.0.tar.gz" or similar
}
}
