% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_optional_dependencies}
\alias{download_optional_dependencies}
\title{Download all optional Arrow dependencies}
\usage{
download_optional_dependencies(deps_dir)
}
\arguments{
\item{deps_dir}{Directory to save files into. Will be created if necessary.}
}
\value{
TRUE/FALSE for whether the downloads were successful

This function is used for setting up an offline build. If it's possible to
download at build time, don't use this function. Instead, let \code{cmake}
download them for you.
These saved files are only used in the build if \code{ARROW_DEPENDENCY_SOURCE}
is \code{BUNDLED} or \code{AUTO}.
https://arrow.apache.org/docs/developers/cpp/building.html#offline-builds
}
\description{
Download all optional Arrow dependencies
}
\examples{
\dontrun{
download_optional_dependencies("arrow-thirdparty")
}
# Now define the environment variables (see arrow-thirdparty/DEFINE_ENV_VARS.sh)
# and run your offline build.
This script is copied over from arrow/cpp/... to arrow/r/tools/cpp/...
}
