% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-arrow.R
\name{download_optional_dependencies}
\alias{download_optional_dependencies}
\title{Download all optional Arrow dependencies}
\usage{
download_optional_dependencies(deps_dir = NULL)
}
\arguments{
\item{deps_dir}{Directory to save files into. Will be created if necessary.
Defaults to the value of \code{ARROW_THIRDPARTY_DEPENDENCY_DIR}, if that
environment variable is set.}
}
\value{
\code{deps_dir}, invisibly

This function is used for setting up an offline build. If it's possible to
download at build time, don't use this function. Instead, let \code{cmake}
download them for you.
If the files already exist in \code{deps_dir}, they will be re-downloaded and
overwritten. Do not put other files in this directory.
These saved files are only used in the build if \code{ARROW_DEPENDENCY_SOURCE}
is \code{BUNDLED} or \code{AUTO}.
https://arrow.apache.org/docs/developers/cpp/building.html#offline-builds
\subsection{Steps for an offline install with optional dependencies:}{
\subsection{On a computer with internet access:}{
\itemize{
\item Install the \code{arrow} package
\item Run this function
\item Copy the saved dependency files to the computer with internet access
}
}

\subsection{On the computer without internet access:}{
\itemize{
\item Create a environment variable called \code{ARROW_THIRDPARTY_DEPENDENCY_DIR} that
points to the newly copied folder of dependency files.
\item Install the \code{arrow} package
\item Run \code{\link[=arrow_info]{arrow_info()}} to check installed capabilities
}
}

}
}
\description{
Download all optional Arrow dependencies
}
\examples{
\dontrun{
download_optional_dependencies("arrow-thirdparty")
list.files("arrow-thirdparty", "thrift-*") # "thrift-0.13.0.tar.gz" or similar
}
}
