% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{enums}
\alias{enums}
\alias{TimeUnit}
\alias{DateUnit}
\alias{Type}
\alias{StatusCode}
\alias{FileMode}
\alias{MessageType}
\alias{CompressionType}
\alias{FileType}
\alias{ParquetVersionType}
\alias{MetadataVersion}
\title{Arrow enums}
\format{
An object of class \code{TimeUnit::type} (inherits from \code{arrow-enum}) of length 4.

An object of class \code{DateUnit} (inherits from \code{arrow-enum}) of length 2.

An object of class \code{Type::type} (inherits from \code{arrow-enum}) of length 36.

An object of class \code{StatusCode} (inherits from \code{arrow-enum}) of length 17.

An object of class \code{FileMode} (inherits from \code{arrow-enum}) of length 3.

An object of class \code{MessageType} (inherits from \code{arrow-enum}) of length 5.

An object of class \code{Compression::type} (inherits from \code{arrow-enum}) of length 9.

An object of class \code{FileType} (inherits from \code{arrow-enum}) of length 4.

An object of class \code{ParquetVersionType} (inherits from \code{arrow-enum}) of length 2.

An object of class \code{MetadataVersion} (inherits from \code{arrow-enum}) of length 5.
}
\usage{
TimeUnit

DateUnit

Type

StatusCode

FileMode

MessageType

CompressionType

FileType

ParquetVersionType

MetadataVersion
}
\description{
Arrow enums
}
\keyword{datasets}
\keyword{internal}
