% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{enums}
\alias{enums}
\alias{TimeUnit}
\alias{DateUnit}
\alias{Type}
\alias{StatusCode}
\alias{FileMode}
\alias{MessageType}
\alias{CompressionType}
\title{Arrow enums}
\format{An object of class \code{arrow::TimeUnit::type} (inherits from \code{arrow-enum}) of length 4.}
\usage{
TimeUnit

DateUnit

Type

StatusCode

FileMode

MessageType

CompressionType
}
\description{
Arrow enums
}
\keyword{datasets}
\keyword{internal}
