% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{FeatherTableReader}
\alias{FeatherTableReader}
\title{A \code{arrow::ipc::feather::TableReader} to read from a file}
\usage{
FeatherTableReader(file, mmap = TRUE, ...)
}
\arguments{
\item{file}{A file path or \code{arrow::io::RandomAccessFile}}

\item{mmap}{Is the file memory mapped (applicable to the \code{character} method)}

\item{...}{extra parameters}
}
\description{
A \code{arrow::ipc::feather::TableReader} to read from a file
}
