% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{FeatherTableWriter}
\alias{FeatherTableWriter}
\title{Create TableWriter that writes into a stream}
\usage{
FeatherTableWriter(stream)
}
\arguments{
\item{stream}{an OutputStream}
}
\description{
Create TableWriter that writes into a stream
}
