% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\docType{class}
\name{FeatherTableWriter}
\alias{FeatherTableWriter}
\title{FeatherTableWriter class}
\description{
This class enables you to write Feather files. See its usage in
\code{\link[=write_feather]{write_feather()}}.
}
\section{Factory}{


The \code{FeatherTableWriter$create()} factory method instantiates the object and
takes the following argument:
\itemize{
\item \code{stream} An \code{OutputStream}
}
}

\section{Methods}{

\itemize{
\item \code{$GetDescription()}
\item \code{$HasDescription()}
\item \code{$version()}
\item \code{$num_rows()}
\item \code{$num_columns()}
\item \code{$GetColumnName()}
\item \code{$GetColumn()}
\item \code{$Read(columns)}
}
}

\keyword{datasets}
