% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{Field}
\alias{Field}
\alias{field}
\title{class arrow::Field}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Field

field(name, type, metadata)
}
\arguments{
\item{name}{field name}

\item{type}{logical type, instance of \link{DataType}}

\item{metadata}{currently ignored}
}
\description{
\code{field()} lets you create an \code{arrow::Field} that maps a
\link[=data-type]{DataType} to a column name. Fields are contained in
\link[=Schema]{Schemas}.
}
\section{Methods}{

\itemize{
\item \code{f$ToString()}: convert to a string
\item \code{f$Equals(other)}: test for equality. More naturally called as \code{f == other}
}
}

\examples{
\donttest{
field("x", int32())
}
}
\keyword{datasets}
