% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{FixedSizeBufferWriter}
\alias{FixedSizeBufferWriter}
\title{Open a \link[=arrow__io__FixedSizeBufferWriter]{arrow::io::FixedSizeBufferWriter}}
\usage{
FixedSizeBufferWriter(buffer)
}
\arguments{
\item{buffer}{\link[=arrow__Buffer]{arrow::Buffer} or something \code{\link[=buffer]{buffer()}} can handle}
}
\value{
a \link[=arrow__io__BufferOutputStream]{arrow::io::BufferOutputStream}
}
\description{
Open a \link[=arrow__io__FixedSizeBufferWriter]{arrow::io::FixedSizeBufferWriter}
}
