% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_get}
\alias{flight_get}
\title{Get data from a Flight server}
\usage{
flight_get(client, path)
}
\arguments{
\item{client}{\code{pyarrow.flight.FlightClient}, as returned by \code{\link[=flight_connect]{flight_connect()}}}

\item{path}{string identifier under which data is stored}
}
\value{
A \link{Table}
}
\description{
Get data from a Flight server
}
