% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_put}
\alias{flight_put}
\title{Send data to a Flight server}
\usage{
flight_put(client, data, path, overwrite = TRUE)
}
\arguments{
\item{client}{\code{pyarrow.flight.FlightClient}, as returned by \code{\link[=flight_connect]{flight_connect()}}}

\item{data}{\code{data.frame}, \link{RecordBatch}, or \link{Table} to upload}

\item{path}{string identifier to store the data under}

\item{overwrite}{logical: if \code{path} exists on \code{client} already, should we
replace it with the contents of \code{data}? Default is \code{TRUE}; if \code{FALSE} and
\code{path} exists, the function will error.}
}
\value{
\code{client}, invisibly.
}
\description{
Send data to a Flight server
}
