% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{hive_partition}
\alias{hive_partition}
\title{Construct a Hive partition scheme}
\usage{
hive_partition(...)
}
\arguments{
\item{...}{named list of \link[=data-type]{data types}, passed to \code{\link[=schema]{schema()}}}
}
\value{
A \code{HivePartitionScheme}
}
\description{
Hive partitioning embeds field names and values in path segments, such as
"/year=2019/month=2/data.parquet". A \link[=PartitionScheme]{HivePartitionScheme}
is used to parse that in Dataset creation.
}
\details{
Because fields are named in the path segments, order of fields passed to
\code{hive_partition()} does not matter.
}
\examples{
\donttest{
hive_partition(year = int16(), month = int8())
}
}
