% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{make_readable_file}
\alias{make_readable_file}
\title{Handle a range of possible input sources}
\usage{
make_readable_file(file, mmap = TRUE, random_access = TRUE)
}
\arguments{
\item{file}{A character file name, \code{raw} vector, or an Arrow input stream}

\item{mmap}{Logical: whether to memory-map the file (default \code{TRUE})}

\item{random_access}{Logical: whether the result must be a RandomAccessFile}
}
\value{
An \code{InputStream} or a subclass of one.
}
\description{
Handle a range of possible input sources
}
\keyword{internal}
