% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{match_arrow}
\alias{match_arrow}
\title{\code{match} for Arrow objects}
\usage{
match_arrow(x, table, ...)
}
\arguments{
\item{x}{\code{Array} or \code{ChunkedArray}}

\item{table}{\code{Array}, \code{ChunkedArray}, or R vector lookup table.}

\item{...}{additional arguments, ignored}
}
\value{
An \code{int32}-type \code{Array} of the same length as \code{x} with the
(0-based) indexes into \code{table}.
}
\description{
\code{base::match()} is not a generic, so we can't just define Arrow methods for
it. This function exposes the analogous function in the Arrow C++ library.
}
