% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\docType{class}
\name{Message}
\alias{Message}
\title{Message}
\description{
Message
}
\section{Methods}{

\itemize{
\item \verb{$Equals(other)}: Test for equality
\item \verb{$body_length()}:
\item \verb{$Verify()}:
}
}

\section{Active bindings}{

\itemize{
\item \verb{$type}:
\item \verb{$metadata}:
\item \verb{$body}:
}
}

