% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{MessageReader}
\alias{MessageReader}
\title{Open a MessageReader that reads from a stream}
\usage{
MessageReader(stream)
}
\arguments{
\item{stream}{an InputStream}
}
\description{
Open a MessageReader that reads from a stream
}
