% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-type-decimal128.R
\name{new_decimal128}
\alias{new_decimal128}
\alias{decimal128}
\alias{format.arrow_decimal128}
\alias{vec_type2.arrow_decimal128}
\alias{vec_cast.arrow_decimal128}
\title{decimal type}
\usage{
new_decimal128(data = complex(), precision = 10L, scale = 0L)

decimal128(data, precision = 10L, scale = 0L)

\method{format}{arrow_decimal128}(x, ...)

\method{vec_type2}{arrow_decimal128}(x, y)

\method{vec_cast}{arrow_decimal128}(x, to)
}
\arguments{
\item{data}{decimal data, hosted in a complex vector}

\item{scale}{scale}

\item{x}{a \code{arrow_decimal128}}
}
\description{
decimal type
}
