% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\docType{class}
\name{ParquetReaderProperties}
\alias{ParquetReaderProperties}
\title{ParquetReaderProperties class}
\description{
This class holds settings to control how a Parquet file is read
by \link{ParquetFileReader}.
}
\section{Factory}{


The \code{ParquetReaderProperties$create()} factory method instantiates the object
and takes no arguments.
}

\section{Methods}{

\itemize{
\item \verb{$thrift_string_size_limit()}
\item \verb{$set_thrift_string_size_limit()}
\item \verb{$thrift_container_size_limit()}
\item \verb{$set_thrift_container_size_limit()}
}
}

