% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\docType{class}
\name{ParquetReaderProperties}
\alias{ParquetReaderProperties}
\title{ParquetReaderProperties class}
\description{
This class holds settings to control how a Parquet file is read
by \link{ParquetFileReader}.
}
\section{Factory}{


The \code{ParquetReaderProperties$create()} factory method instantiates the object
and takes the following arguments:
\itemize{
\item \code{use_threads} Logical: whether to use multithreading (default \code{TRUE})
}
}

\section{Methods}{

\itemize{
\item \code{$read_dictionary(column_index)}
\item \code{$set_read_dictionary(column_index, read_dict)}
\item \code{$use_threads(use_threads)}
}
}

\keyword{datasets}
