% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{read_arrow}
\alias{read_arrow}
\title{Read an tibble from an arrow::Table on disk}
\usage{
read_arrow(path)
}
\arguments{
\item{path}{binary arrow file}
}
\value{
a \link[tibble:tibble]{tibble::tibble}
}
\description{
Read an tibble from an arrow::Table on disk
}
