% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_csv_arrow}
\alias{read_csv_arrow}
\title{Read csv file into an arrow::Table}
\usage{
read_csv_arrow(...)
}
\arguments{
\item{...}{Used to construct an arrow::csv::TableReader}
}
\description{
Use arrow::csv::TableReader from \code{\link[=csv_table_reader]{csv_table_reader()}}
}
