% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{read_feather}
\alias{read_feather}
\title{Read a feather file}
\usage{
read_feather(reader)
}
\arguments{
\item{reader}{a arrow::ipc::feather::TableReader or whatever the \code{\link[=table_reader]{table_reader()}} function can handle}
}
\value{
an arrow::Table
}
\description{
Read a feather file
}
