% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{read_feather}
\alias{read_feather}
\title{Read a feather file}
\usage{
read_feather(file, ...)
}
\arguments{
\item{file}{a arrow::ipc::feather::TableReader or whatever the \code{\link[=feather_table_reader]{feather_table_reader()}} function can handle}

\item{...}{additional parameters}
}
\value{
an arrow::Table
}
\description{
Read a feather file
}
