% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{read_feather}
\alias{read_feather}
\title{Read a feather file}
\usage{
read_feather(file, columns = NULL, as_tibble = TRUE,
  use_threads = TRUE, ...)
}
\arguments{
\item{file}{a arrow::ipc::feather::TableReader or whatever the \code{\link[=FeatherTableReader]{FeatherTableReader()}} function can handle}

\item{columns}{names if the columns to read. The default \code{NULL} means all columns}

\item{as_tibble}{should the \link[=arrow__Table]{arrow::Table} be converted to a tibble.}

\item{use_threads}{Use threads when converting to a tibble.}

\item{...}{additional parameters}
}
\value{
a data frame if \code{as_tibble} is \code{TRUE} (the default), or a \link[=arrow__Table]{arrow::Table} otherwise
}
\description{
Read a feather file
}
