% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read a Parquet file}
\usage{
read_parquet(file, col_select = NULL, as_tibble = TRUE,
  props = ParquetReaderProperties$create(), ...)
}
\arguments{
\item{file}{A character file name, raw vector, or an Arrow input stream}

\item{col_select}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:vars_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}

\item{as_tibble}{Should the function return a \code{data.frame} or an
\link[=Table]{arrow::Table}?}

\item{props}{\link{ParquetReaderProperties}}

\item{...}{Additional arguments passed to \code{ParquetFileReader$create()}}
}
\value{
A \link[=Table]{arrow::Table}, or a \code{data.frame} if \code{as_tibble} is
\code{TRUE}.
}
\description{
'\href{https://parquet.apache.org/}{Parquet}' is a columnar storage file format.
This function enables you to read Parquet files into R.
}
\examples{
\donttest{
df <- read_parquet(system.file("v0.7.1.parquet", package="arrow"))
head(df)
}
}
