% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read parquet file from disk}
\usage{
read_parquet(file, as_tibble = TRUE, use_threads = TRUE, ...)
}
\arguments{
\item{file}{a file path}

\item{as_tibble}{should the \link[=arrow__Table]{arrow::Table} be converted to a tibble.}

\item{use_threads}{Use threads when converting to a tibble, only relevant if \code{as_tibble} is \code{TRUE}}

\item{...}{currently ignored}
}
\value{
a \link[=arrow__Table]{arrow::Table}, or a data frame if \code{as_tibble} is \code{TRUE}.
}
\description{
Read parquet file from disk
}
