% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read Parquet file from disk}
\usage{
read_parquet(file, as_tibble = TRUE, ...)
}
\arguments{
\item{file}{a file path}

\item{as_tibble}{Should the \link[=arrow__Table]{arrow::Table} be converted to a
tibble? Default is \code{TRUE}.}

\item{...}{Additional arguments, currently ignored}
}
\value{
A \link[=arrow__Table]{arrow::Table}, or a \code{tbl_df} if \code{as_tibble} is
\code{TRUE}.
}
\description{
'\href{https://parquet.apache.org/}{Parquet}' is a columnar storage file format.
This function enables you to read Parquet files into R.
}
\examples{

\dontrun{
  df <- read_parquet(system.file("v0.7.1.parquet", package="arrow"))
}

}
