% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read_record_batch}
\alias{read_record_batch}
\title{Read a \link{RecordBatch} from an encapsulated IPC message}
\usage{
read_record_batch(obj, schema)
}
\arguments{
\item{obj}{a \link{Message}, a \link{InputStream}, a \link{Buffer}, or a \code{raw} vector}

\item{schema}{a \link{schema}}
}
\value{
a \link{RecordBatch}
}
\description{
Deprecated in 0.17. Use \code{\link[=record_batch]{record_batch()}} instead.
}
\keyword{internal}
