% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatch.R
\name{read_record_batch}
\alias{read_record_batch}
\title{Read a single record batch from a stream}
\usage{
read_record_batch(stream)
}
\arguments{
\item{stream}{input stream}
}
\description{
Read a single record batch from a stream
}
\details{
\code{stream} can be a \code{arrow::io::RandomAccessFile} stream as created by \code{\link[=file_open]{file_open()}} or \code{\link[=mmap_open]{mmap_open()}} or a path.
}
