% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_record_batch.R
\name{read_record_batch}
\alias{read_record_batch}
\title{read \link[=arrow__RecordBatch]{arrow::RecordBatch} as encapsulated IPC message, given a known \link[=arrow__Schema]{arrow::Schema}}
\usage{
read_record_batch(obj, schema)
}
\arguments{
\item{obj}{a \link[=arrow__ipc__Message]{arrow::ipc::Message}, a \link[=arrow__io__InputStream]{arrow::io::InputStream}, a \link[=arrow__Buffer]{arrow::Buffer}, or a raw vector}

\item{schema}{a \link[=arrow__Schema]{arrow::Schema}}
}
\value{
a \link[=arrow__RecordBatch]{arrow::RecordBatch}
}
\description{
read \link[=arrow__RecordBatch]{arrow::RecordBatch} as encapsulated IPC message, given a known \link[=arrow__Schema]{arrow::Schema}
}
