% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-record-batch.R
\name{read_record_batch}
\alias{read_record_batch}
\title{read \link[=RecordBatch]{arrow::RecordBatch} as encapsulated IPC message, given a known \link[=schema]{arrow::Schema}}
\usage{
read_record_batch(obj, schema)
}
\arguments{
\item{obj}{a \link[=Message]{arrow::Message}, a \link[=InputStream]{arrow::io::InputStream}, a \link[=buffer]{Buffer}, or a raw vector}

\item{schema}{a \link[=schema]{arrow::Schema}}
}
\value{
a \link[=RecordBatch]{arrow::RecordBatch}
}
\description{
read \link[=RecordBatch]{arrow::RecordBatch} as encapsulated IPC message, given a known \link[=schema]{arrow::Schema}
}
