% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{read_schema}
\alias{read_schema}
\title{read a Schema from a stream}
\usage{
read_schema(stream, ...)
}
\arguments{
\item{stream}{a stream}

\item{...}{currently ignored}
}
\description{
read a Schema from a stream
}
