% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-batch.R
\name{record_batch}
\alias{record_batch}
\title{Create an \link[=RecordBatch]{arrow::RecordBatch} from a data frame}
\usage{
record_batch(..., schema = NULL)
}
\arguments{
\item{...}{A variable number of Array}

\item{schema}{a arrow::Schema}
}
\value{
a \link[=RecordBatch]{arrow::RecordBatch}
}
\description{
Create an \link[=RecordBatch]{arrow::RecordBatch} from a data frame
}
