% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatch.R
\name{record_batch}
\alias{record_batch}
\title{Create an \link[=arrow__RecordBatch]{arrow::RecordBatch} from a data frame}
\usage{
record_batch(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a \link[=arrow__RecordBatch]{arrow::RecordBatch}
}
\description{
Create an \link[=arrow__RecordBatch]{arrow::RecordBatch} from a data frame
}
