% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\name{RecordBatchFileReader}
\alias{RecordBatchFileReader}
\title{Create an \link[=arrow__ipc__RecordBatchFileReader]{arrow::ipc::RecordBatchFileReader} from a file}
\usage{
RecordBatchFileReader(file)
}
\arguments{
\item{file}{The file to read from. A file path, or an \link[=arrow__ipc__RecordBatchFileReader]{arrow::io::RandomAccessFile}}
}
\description{
Create an \link[=arrow__ipc__RecordBatchFileReader]{arrow::ipc::RecordBatchFileReader} from a file
}
