% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{RecordBatchFileWriter}
\alias{RecordBatchFileWriter}
\title{Create a record batch file writer from a stream}
\usage{
RecordBatchFileWriter(sink, schema)
}
\arguments{
\item{sink}{Where to write. Can either be:
\itemize{
\item a string file path
\item \link[=arrow__io__OutputStream]{arrow::io::OutputStream}
}}

\item{schema}{The \link[=arrow__Schema]{arrow::Schema} for data to be written.}
}
\value{
an \code{arrow::ipc::RecordBatchWriter} object
}
\description{
Create a record batch file writer from a stream
}
