% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchReader.R
\name{RecordBatchStreamReader}
\alias{RecordBatchStreamReader}
\title{Create a \link[=arrow__ipc__RecordBatchStreamReader]{arrow::ipc::RecordBatchStreamReader} from an input stream}
\usage{
RecordBatchStreamReader(stream)
}
\arguments{
\item{stream}{input stream, an \link[=arrow__io__InputStream]{arrow::io::InputStream} or a raw vector}
}
\description{
Create a \link[=arrow__ipc__RecordBatchStreamReader]{arrow::ipc::RecordBatchStreamReader} from an input stream
}
