% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{RecordBatchStreamWriter}
\alias{RecordBatchStreamWriter}
\title{Writer for the Arrow streaming binary format}
\usage{
RecordBatchStreamWriter(sink, schema)
}
\arguments{
\item{sink}{Where to write. Can either be:
\itemize{
\item A string, meant as a file path, passed to \code{\link[fs:path_abs]{fs::path_abs()}}
\item a \link[fs:path_abs]{file path}
\item \link[=arrow__io__OutputStream]{arrow::io::OutputStream}
}}

\item{schema}{The \link[=arrow__Schema]{arrow::Schema} for data to be written.}
}
\value{
a \link[=arrow__ipc__RecordBatchStreamWriter]{arrow::ipc::RecordBatchStreamWriter}
}
\description{
Writer for the Arrow streaming binary format
}
