% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-bit64.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{print.integer64}
\alias{str.integer64}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bit64}{\code{\link[bit64]{print.integer64}}, \code{\link[bit64]{str.integer64}}}
}}

