% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{s3_init}
\alias{s3_init}
\title{Initialize S3}
\usage{
s3_init(log_level)
}
\arguments{
\item{log_level}{string for log level to set. See details.}
}
\description{
Must be called before other S3 functions such as \link{s3_bucket}.
}
\details{
The parameter \code{log_level} must be one of: Off, Fatal, Error, Warn,
Info, Debug, or Trace.
}
\examples{
# Initialize S3 with Debug-level logging
\dontrun{
  s3_init("Debug")
}
}
