% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\docType{class}
\name{Scalar}
\alias{Scalar}
\title{Arrow scalars}
\description{
A \code{Scalar} holds a single value of an Arrow type.
}
\section{Methods}{

\verb{$ToString()}: convert to a string
\verb{$as_vector()}: convert to an R vector
\verb{$as_array()}: convert to an Arrow \code{Array}
\verb{$Equals(other)}: is this Scalar equal to \code{other}
\verb{$ApproxEquals(other)}: is this Scalar approximately equal to \code{other}
\verb{$is_valid}: is this Scalar valid
\verb{$null_count}: number of invalid values - 1 or 0
\verb{$type}: Scalar type
}

