% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{scalar}
\alias{scalar}
\alias{StructScalar}
\title{Create an Arrow Scalar}
\usage{
scalar(x, type = NULL)
}
\arguments{
\item{x}{An R vector, list, or \code{data.frame}}

\item{type}{An optional \link[=data-type]{data type} for \code{x}. If omitted, the type will be inferred from the data.}
}
\description{
Create an Arrow Scalar
}
\examples{
scalar(pi)
scalar(404)
# If you pass a vector into scalar(), you get a list containing your items
scalar(c(1, 2, 3))

scalar(9) == scalar(10)

}
