% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{class}
\name{Schema}
\alias{Schema}
\alias{schema}
\title{Schema class}
\usage{
schema(...)
}
\arguments{
\item{...}{named list of \link[=data-type]{data types}}
}
\description{
Create a \code{Schema} when you
want to convert an R \code{data.frame} to Arrow but don't want to rely on the
default mapping of R types to Arrow types, such as when you want to choose a
specific numeric precision.
}
\section{Usage}{
\preformatted{s <- schema(...)

s$ToString()
s$num_fields()
s$field(i)
}
}

\section{Methods}{

\itemize{
\item \code{$ToString()}: convert to a string
\item \code{$num_fields()}: returns the number of fields
\item \code{$field(i)}: returns the field at index \code{i} (0-based)
}
}

\keyword{datasets}
