% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Source}
\alias{Source}
\alias{SourceDiscovery}
\alias{FileSystemSourceDiscovery}
\title{Data sources for a Dataset}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Source
}
\description{
A \link{Dataset} can have one or more \code{Source}s. A \code{Source} contains one
or more \code{Fragments}, such as files, of a common type and partition
scheme. \code{SourceDiscovery} is used to create a \code{Source}, inspect the
\link{Schema} of the fragments contained in it, and declare a partition scheme.
\code{FileSystemSourceDiscovery} is a subclass of \code{SourceDiscovery} for
discovering files in the local file system, the only currently supported
file system.
}
\section{Factory}{

The \code{SourceDiscovery$create()} factory method instantiates a
\code{SourceDiscovery} and takes the following arguments:
\itemize{
\item \code{path}: A string file path containing data files
\item \code{filesystem}: Currently only "local" is supported
\item \code{format}: Currently only "parquet" is supported
\item \code{allow_non_existent}: logical: is \code{path} allowed to not exist? Default
\code{FALSE}. See \link{FileSelector}.
\item \code{recursive}: logical: should files be discovered in subdirectories of
\item \code{path}? Default \code{TRUE}.
\item \code{...} Additional arguments passed to the \link{FileSystem} \code{$create()} method
}

\code{FileSystemSourceDiscovery$create()} is a lower-level factory method and
takes the following arguments:
\itemize{
\item \code{filesystem}: A \link{FileSystem}
\item \code{selector}: A \link{FileSelector}
\item \code{format}: Currently only "parquet" is supported
}
}

\section{Methods}{

\code{Source} has no defined methods. It is just passed to \code{Dataset$create()}.

\code{SourceDiscovery} and its subclasses have the following methods:
\itemize{
\item \code{$Inspect()}: Walks the files in the directory and returns a common \link{Schema}
\item \code{$Finish(schema)}: Returns a \code{Source}
}
}

\seealso{
\link{Dataset} for what do do with a \code{Source}
}
\keyword{datasets}
