% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{table}
\alias{table}
\title{Create an arrow::Table from a data frame}
\usage{
table(..., schema = NULL)
}
\arguments{
\item{...}{arrays, chunked arrays, or R vectors}

\item{schema}{NULL or a schema}
}
\description{
Create an arrow::Table from a data frame
}
