% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{table}
\alias{table}
\title{Create an arrow::Table from a data frame}
\usage{
table(..., schema = NULL)
}
\arguments{
\item{...}{arrays, chunked arrays, or R vectors}

\item{schema}{a schema. The default (\code{NULL}) infers the schema from the \code{...}}
}
\value{
an arrow::Table
}
\description{
Create an arrow::Table from a data frame
}
