% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{class}
\name{Table}
\alias{Table}
\title{class arrow::Table}
\description{
class arrow::Table
}
\section{Factory}{


The \code{Table$create()} function takes the following arguments:
\itemize{
\item \code{...} arrays, chunked arrays, or R vectors
\item \code{schema} a schema. The default (\code{NULL}) infers the schema from the \code{...}
}
}

\section{Methods}{


TODO
}

\keyword{datasets}
