% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type}
\alias{type}
\title{infer the arrow Array type from an R vector}
\usage{
type(x)
}
\arguments{
\item{x}{an R vector}
}
\value{
an arrow logical type
}
\description{
infer the arrow Array type from an R vector
}
