% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_arrow.R
\name{write_arrow}
\alias{write_arrow}
\title{Write Arrow formatted data}
\usage{
write_arrow(x, stream, ...)
}
\arguments{
\item{x}{an \link[=arrow__Table]{arrow::Table}, an \link[=arrow__RecordBatch]{arrow::RecordBatch} or a data.frame}

\item{stream}{where to serialize to
\itemize{
\item A \link[=arrow__ipc__RecordBatchWriter]{arrow::ipc::RecordBatchWriter}: the \code{$write()}
of \code{x} is used. The stream is left open. This uses the streaming format
or the binary file format depending on the type of the writer.
\item A string file path: \code{x} is serialized with
a \link[=arrow__ipc__RecordBatchFileWriter]{arrow::ipc::RecordBatchFileWriter}, i.e.
using the binary file format.
\item A raw vector: typically of length zero (its data is ignored, and only used for
dispatch). \code{x} is serialized using the streaming format, i.e. using the
\link[=arrow__ipc__RecordBatchStreamWriter]{arrow::ipc::RecordBatchStreamWriter}
}}

\item{...}{extra parameters, currently ignored

\code{write_arrow} is a convenience function, the classes \link[=arrow__ipc__RecordBatchFileWriter]{arrow::ipc::RecordBatchFileWriter}
and \link[=arrow__ipc__RecordBatchStreamWriter]{arrow::ipc::RecordBatchStreamWriter} can be used for more flexibility.}
}
\description{
Write Arrow formatted data
}
