% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordBatchWriter.R
\name{write_arrow}
\alias{write_arrow}
\title{Write an object to a stream}
\usage{
write_arrow(x, stream, ...)
}
\arguments{
\item{x}{An object to stream}

\item{stream}{A stream}

\item{...}{additional parameters}
}
\description{
Write an object to a stream
}
