% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{write_arrow}
\alias{write_arrow}
\title{Write a tibble in a binary arrow file}
\usage{
write_arrow(data, path)
}
\arguments{
\item{data}{a \link[tibble:tibble]{tibble::tibble}}

\item{path}{file path}
}
\description{
Write a tibble in a binary arrow file
}
