% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-package.R
\name{write_arrow}
\alias{write_arrow}
\title{Use \code{write_ipc_stream} or \code{write_to_raw} instead.}
\usage{
write_arrow(...)
}
\arguments{
\item{...}{Ignored arguments}
}
\description{
Use \code{write_ipc_stream} or \code{write_to_raw} instead.
}
\keyword{internal}
