% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{write_csv_arrow}
\alias{write_csv_arrow}
\title{Write CSV file to disk}
\usage{
write_csv_arrow(x, sink, include_header = TRUE, batch_size = 1024L)
}
\arguments{
\item{x}{\code{data.frame}, \link{RecordBatch}, or \link{Table}}

\item{sink}{A string file path, URI, or \link{OutputStream}, or path in a file
system (\code{SubTreeFileSystem})}

\item{include_header}{Whether to write an initial header line with column names}

\item{batch_size}{Maximum number of rows processed at a time. Default is 1024}
}
\value{
The input \code{x}, invisibly. Note that if \code{sink} is an \link{OutputStream},
the stream will be left open.
}
\description{
Write CSV file to disk
}
\examples{
\donttest{
tf <- tempfile()
on.exit(unlink(tf))
write_csv_arrow(mtcars, tf)
}
}
