% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-write.R
\name{write_dataset}
\alias{write_dataset}
\title{Write a dataset}
\usage{
write_dataset(
  dataset,
  path,
  format = dataset$format$type,
  partitioning = dplyr::group_vars(dataset),
  hive_style = TRUE,
  ...
)
}
\arguments{
\item{dataset}{\link{Dataset} or \code{arrow_dplyr_query}. If a \code{arrow_dplyr_query},
note that \code{select()} or \code{filter()} queries are not currently supported.}

\item{path}{string path to a directory to write to (directory will be
created if it does not exist)}

\item{format}{file format to write the dataset to. Currently only "feather"
(aka "ipc") is supported.}

\item{partitioning}{\code{Partitioning} or a character vector of columns to
use as partition keys (to be written as path segments). Default is to
use the current \code{group_by()} columns.}

\item{hive_style}{logical: write partition segments as Hive-style
(\code{key1=value1/key2=value2/file.ext}) or as just bare values. Default is \code{TRUE}.}

\item{...}{additional arguments, passed to \code{dataset$write()}}
}
\value{
The input \code{dataset}, invisibly
}
\description{
This function allows you to write a dataset. By writing to more efficient
binary storage formats, and by specifying relevant partitioning, you can
make it much faster to read and query.
}
