% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{write_feather}
\alias{write_feather}
\title{Write data in the Feather format}
\usage{
write_feather(x, sink)
}
\arguments{
\item{x}{\code{data.frame} or RecordBatch}

\item{sink}{A file path or an OutputStream}
}
\value{
the input \code{x} invisibly.
}
\description{
Write data in the Feather format
}
\examples{
\donttest{
tf <- tempfile()
on.exit(unlink(tf))
write_feather(mtcars, tf)
}
}
