% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{write_feather}
\alias{write_feather}
\title{Write data in the Feather format}
\usage{
write_feather(data, stream)
}
\arguments{
\item{data}{\code{data.frame} or RecordBatch}

\item{stream}{A file path or an OutputStream}
}
\description{
Write data in the Feather format
}
\examples{
\donttest{
try({
  tf <- tempfile()
  on.exit(unlink(tf))
  write_feather(mtcars, tf)
})
}
}
