% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{write_feather}
\alias{write_feather}
\alias{write_feather_RecordBatch}
\title{Write data in the feather format}
\usage{
write_feather(data, stream)

write_feather_RecordBatch(data, stream)
}
\arguments{
\item{data}{frame or arrow::RecordBatch}

\item{stream}{A file path or an arrow::io::OutputStream}
}
\description{
Write data in the feather format
}
