% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{write_parquet}
\alias{write_parquet}
\title{Write Parquet file to disk}
\usage{
write_parquet(table, stream)
}
\arguments{
\item{table}{An \link[=Table]{arrow::Table}, or an object convertible to it with \code{\link[=to_arrow]{to_arrow()}}}

\item{stream}{an \link[=OutputStream]{arrow::io::OutputStream} or a string which is interpreted as a file path}
}
\description{
\href{https://parquet.apache.org/}{Parquet} is a columnar storage file format.
This function enables you to write Parquet files from R.
}
\examples{
\donttest{
tf1 <- tempfile(fileext = ".parquet")
write_parquet(tibble::tibble(x = 1:5), tf2)

# using compression
tf2 <- tempfile(fileext = ".gz.parquet")
write_parquet(tibble::tibble(x = 1:5), CompressedOutputStream$create(tf2, "gzip"))

}
}
